% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{add.abundance}
\alias{add.abundance}
\title{Add abundance to sample data}
\usage{
add.abundance(sdata, ..., phy, counts = FALSE)
}
\arguments{
\item{sdata}{sample data to be modified}

\item{...}{one or more taxonomic expressions defining the abundance to be calculated. Can be named, in order to specify the column name.}

\item{phy}{phyloseq object to be used for caluclation}

\item{counts}{if \code{TRUE}, will return count rather than relative abundance}
}
\value{
returns \code{sdata}, with additional columns for abundance.
}
\description{
Add abundance to sample data
}
\examples{
get.samp(cid.phy) \%>\% add.abundance(pct.entero=Genus=="Enterococcus",pct.proteo=Phylum=="Proteobacteria",phy=cid.phy)

}
