% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{add.code.dependencies}
\alias{add.code.dependencies}
\title{Add Code Dependencies}
\usage{
add.code.dependencies(expr, text = NULL, fn = NULL)
}
\arguments{
\item{expr}{expression to be evaluated.}

\item{text}{character, code can alternatively input as a string.}

\item{fn}{function, use this to evaluate and modify code within a function.}
}
\value{
returns either a modified expression, text, or function, depending on what was input.
}
\description{
Given code (as an expression, text, or function), examine for package dependencies and
load them.
}
\examples{
add.code.dependencies(overlaps(1,2,3,sqrt(44)))

add.code.dependencies(text="log_epsilon_trans(0.001)")

fun <- function() {
  age.years(as.Date("1975-02-21"),Sys.Date())
}
add.code.dependencies(fn=fun)
}
