% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{all.grepl}
\alias{all.grepl}
\title{All Grepl}
\usage{
\method{all}{grepl}(pattern, x, n.screen = 10000, ...)
}
\description{
Equivalent to \code{all(grepl(...))}. Performs an initial screen of first 10000 values, to save time.
}
\details{
This is used in \code{as.Date2} function to save time on pattern matching.
}
