% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{all.same}
\alias{all.same}
\alias{all.same.default}
\alias{all.same.data.frame}
\title{Test if 2 objects are equal}
\usage{
all.same(x, ...)

\method{all}{same.default}(x, y)

\method{all}{same.data.frame}(x, y)
}
\arguments{
\item{x}{first object}

\item{y}{second object}
}
\value{
logical vector
}
\description{
Similar to \code{all.equal(x,y)}, but without comments,
and ignores row order in data frames
}
\examples{
x <- c("A","B","C",NA,NA)
y <- c("A","X",NA,"D",NA)
all.same(x,y)
}
