% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{annotation_ruler}
\alias{annotation_ruler}
\title{Annotation Ruler}
\usage{
annotation_ruler(
  units = "cm",
  colour = "black",
  size = 0.5,
  linetype = 1,
  alpha = 1
)
}
\arguments{
\item{units}{Units to use. Default is \code{"cm"}. Can be values from \code{\link[unit:grid]{unit::grid()}}}

\item{colour}{Color to use. Default is \code{"blue"}.}

\item{size}{Line size to use. Default is \code{0.5}.}

\item{linetype}{Line type to use. Default is \code{1}.}

\item{alpha}{Alpha to use. Default is \code{1}.}
}
\description{
Add a ruler to the plot, for sizing, eyeballing purposes.
}
\examples{
library(tidyverse)
g <- ggplot(iris,aes(x=Petal.Length,y=Petal.Width,color=Species,label=Species)) +
  geom_point(size=5,alpha=0.5) +
  geom_text(color="black",check_overlap = TRUE)
g + annotation_ruler("cm")
g + annotation_ruler("inches")
g + annotation_ruler("mm")
}
