% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{any.overlap}
\alias{any.overlap}
\title{Any overlap}
\usage{
\method{any}{overlap}(start, stop, check = TRUE, na.rm = TRUE)
}
\arguments{
\item{start}{vector specifying the start of the intervals}

\item{stop}{vector specifying the end of the intervals}

\item{check}{whether to check if start is greater than stop (default is \code{TRUE})}

\item{na.rm}{whether to remove NA values (default is TRUE)}
}
\value{
whether or not at least one interval is overlapping.
}
\description{
For a given set of intervals, determine whether any interval is overlapping.
}
