% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{as.Date2}
\alias{as.Date2}
\title{Convert to Date (Ying's version)}
\usage{
as.Date2(vec)
}
\arguments{
\item{vec}{The vector to be converted.}
}
\value{
Returns vector converted to Date. If date+time, convert to POSIXct.
}
\description{
Examines strings and convert to Date format if necessary.
}
\examples{
asdf
}

