% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{as.mrn}
\alias{as.mrn}
\alias{as.mrn.default}
\alias{as.mrn.data.frame}
\title{Convert to MRN format}
\usage{
as.mrn(x, ...)

\method{as.mrn}{default}(mrn)

\method{as.mrn}{data.frame}(data, verbose = FALSE)
}
\arguments{
\item{data}{Character or numeric vector or data frame to be analyzed.}
}
\value{
An 8-character vector is returned. If data frame is entered, a data frame with MRNs converted will be returned.
}
\description{
Formats data containing MRNs
}
\details{
Formats MRN data. When data is pulled in from other sources (e.g. Excel, Access),
the MRNs may be formatted differently depending on source. This unifies the formatting
of MRNs by converting to a 8-character vector. Adds 0's to the beginning if necessary,
to make every value 8 characters long. If a data frame is entered, the same data frame
is returned, but any column resembling an MRN will be formatted (if column name is not
already "MRN", an extra column will be added to the data frame).
}
\examples{
as.mrn(1234)
as.mrn("1234")
}
\author{
Ying Taur
}
