% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{as.phylo.formula2}
\alias{as.phylo.formula2}
\title{Conversion from Taxonomy Variables to Phylogenetic Trees (YT converted)}
\usage{
as.phylo.formula2(
  x,
  data = parent.frame(),
  collapse.singles = FALSE,
  distinct.tree = TRUE,
  full.taxonomy.only = TRUE,
  ...
)
}
\arguments{
\item{x}{a right-side formula describing the taxonomic relationship: ~C1/C2/.../Cn.}

\item{data}{the data.frame where to look for the variables (default to environment).}

\item{collapse.singles}{whether or not to collapse singleton nodes. Default is FALSE.}

\item{...}{further arguments to be passed from other methods.}
}
\value{
An object of class phylo.
}
\description{
Used to convert taxonomy table into a phylo object for plotting. A revised version of ape::as.phylo.formula.
Modified from a version from Liam J. Revell, University of Massachusetts, \link{https://stat.ethz.ch/pipermail/r-sig-phylo/2013-August/003017.html}
}
\details{
Here are the changes:
(1) Corrected branch lengths. The original ape::as.phylo.formula does not work well because it uses ape::read.tree to read in data,
which can only read Newick trees without singleton branches. This uses read.newick instead, which can handle the singleton branches.
(2) Single branches are not automatically collapsed. This is so you can plot all levels of taxonomy.
(3) All nodes are labeled.
}
\examples{
#levels are not same level.
library(ape)
data(carnivora)
t1 <- as.phylo.formula(~SuperFamily/Family/Genus/Species, data=carnivora)
par(lend=2)
plot(t1,edge.width=2,cex=0.6,no.margin=TRUE)
#this is correct.
t2 <- as.phylo.formula2(~SuperFamily/Family/Genus/Species, data=carnivora)
par(lend=2)
plot(t2,edge.width=2,cex=0.6,no.margin=TRUE)
}
\author{
Ying Taur
}
