% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{as_date_fractional}
\alias{as_date_fractional}
\title{Convert POSIXct to a fractional date}
\usage{
as_date_fractional(datetime)
}
\arguments{
\item{datetime}{a POSIXct vector to be convered}
}
\value{
a Date column that contains fractional values.
}
\description{
This creates a fractional date. If you do as.Date(datetime), it removes the time part.
}
