% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{bracketGrob}
\alias{bracketGrob}
\title{bracketGrob}
\usage{
bracketGrob(
  x0 = unit(0, "npc"),
  y0 = unit(0, "npc"),
  x1 = unit(1, "npc"),
  y1 = unit(1, "npc"),
  default.units = "npc",
  bracket.width = unit(1, "cm"),
  tip = "round",
  arrow = NULL,
  gp.segment = gpar(),
  label = NA,
  angle = NA,
  hjust = NULL,
  vjust = NULL,
  check.overlap = FALSE,
  gp.text = gpar(),
  name = NULL,
  vp = NULL
)
}
\description{
This creates the brackets in \code{\link[=geom_bracket]{geom_bracket()}}.
}
\examples{
library(tidyverse)
x0=0.15; x1=0.85; y0=0.5; y1=0.5; bw=0.1
ggplot(mtcars) +
  geom_bracket(aes(x=x0,xend=x1,y=y0,yend=y1,label="[Bracket label]"),linewidth=1.25,fontsize=6,bracket.width = unit(bw,"npc")) +
  annotate("segment",x=x0,y=y0,xend=x1,yend=y1,linetype="longdash",color="blue") +
  annotate("point",x=x0,y=y0,color="blue") +
  annotate("text",x=x0,y=y0+0.01,label="(x,y)",color="blue") +
  annotate("point",x=x1,y=y1,color="blue") +
  annotate("text",x=x1,y=y1+0.01,label="(xend,yend)",color="blue") +
  geom_bracket(aes(x=x0-0.35*bw,y=y0,yend=y0-bw,label="bracket.width"),curly=FALSE,color="blue",flip=TRUE) +
  geom_bracket(aes(x=0.6,y=y0,yend=y0+bw,label="bracket.width"),curly=FALSE,color="blue",flip=TRUE) +
  geom_bracket(aes(x=x0,xend=x0+2.5*bw,y=y0-1.25*bw,label="curve starts at\n2.5 * bracket.width"),curly=FALSE,color="blue",flip=TRUE) +
  coord_fixed(xlim=c(0,1),ylim=c(0,1),expand=FALSE) +
  theme_void()
}
