% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{calc.mean.distance}
\alias{calc.mean.distance}
\title{Calculate distance matrix at all taxonomic levels and take weighted average}
\usage{
calc.mean.distance(phy, method, weights = NULL, verbose = TRUE)
}
\arguments{
\item{phy}{phyloseq data}

\item{method}{distance metric method. Passed to \code{\link[=calc.distance]{calc.distance()}}}

\item{weights}{vector of weights to be used when averaging distances.
A value should be provided for each taxonomic level, plus one for OTU-level.
Default (\code{NULL}) is higher levels are weighted higher, and top level is 0.
For example, for 8 levels (\code{Superkingdom}, \code{Phylum}, \code{Class}, \code{Order}, \code{Family}, \code{Genus}, \code{Species}, \code{otu}),
the weights would be \code{c(0,7,6,5,4,3,2,1)}}

\item{verbose}{whether or not to display progress info while calculating.}
}
\value{
a distance matrix representing the averaged distances from \code{phy}
}
\description{
Calculate distance matrix at all taxonomic levels and take weighted average
}
