% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{calc.taxhorn.distance}
\alias{calc.taxhorn.distance}
\title{Calculate the \code{taxhorn} distance}
\usage{
calc.taxhorn.distance(phy)
}
\arguments{
\item{phy}{phyloseq object}
}
\value{
distance metric of \code{taxhorn} distances
}
\description{
Ephraim Slamka helped to develop this metric, in which the Horn distance is calculated over after
collapsing at each taxonomic level and then taking the weighted average of distance values.
}
