% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microbiota2.R
\docType{data}
\name{cid.colors}
\alias{cid.colors}
\title{The color scheme used in CID manuscript.}
\format{\preformatted{ Named chr [1:21] "#129246" "#a89e6a" "#f69ea0" "#2dbfc2" ...
 - attr(*, "names")= chr [1:21] "Enterococcus" "Streptococcus" "Blautia" "Bacteroides" ...
}}
\usage{
cid.colors
}
\description{
The color scheme used in CID manuscript.
}
\author{
Ying Taur
}
\keyword{datasets}

