% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cid94}
\alias{cid94}
\title{CID Data (94 patients)}
\format{Several data frames.}
\usage{
data(cid94)
}
\description{
A dataset containing clinical and microbiota data on the 94 patients from CID 2012.
}
\details{
\itemize{
  \item pt Patient data (N=94)
  \item med Medications dispensed
  \item bsi Bloodstream infection data
  \item cdiff C.difficile testing
  \item fever Tmax for patients
  \item hosp Hospitalizations
}
}
\keyword{datasets}
