% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{cleanup.data}
\alias{cleanup.data}
\title{Cleanup data}
\usage{
cleanup.data(data, make.names = TRUE, trim = TRUE, convert.dates = TRUE,
  as.mrn = TRUE, remove.na.cols.rows = FALSE)
}
\arguments{
\item{make.names}{If \code{TRUE}, will fix variable names. Default=\code{TRUE}}

\item{trim}{If \code{TRUE}, will remove whitespace from all character variables. Default=\code{TRUE}}

\item{convert.dates}{If \code{TRUE}, will convert variables that look like dates to Date format. Default=\code{TRUE}}

\item{as.mrn}{If \code{TRUE}, will looking for variables that look like MRN and convert to 8-digit character. Default=\code{TRUE}}

\item{remove.na.cols.rows}{If \code{TRUE}, will remove any column or row consisting entirely of \code{NA}'s. Default=\code{FALSE}}
}
\value{
Returns a clean version of \code{data}.
}
\description{
Cleans up a data frame by performing 4 tasks:
(1) \code{trim}: for all character variables, remove whitespace
(2) \code{convert.dates}: for all character variables, remove whitespace
(3) \code{as.mrn}: will properly format any MRN variable.
(4) Remove any column or row that is all \code{NA} values.
}
\examples{
#####
}
\author{
Ying Taur
}

