% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{cli_table}
\alias{cli_table}
\title{CLI table}
\usage{
cli_table(data, flexcols = NULL, sep = " ")
}
\arguments{
\item{data}{data frame of CLI text}

\item{flexcols}{vector of column indices that can be truncated to fit the screen width.
Default is \code{NULL}, which selects the widest columns.}

\item{sep}{separator between columns. Default is \code{" "}}
}
\description{
Display table with CLI text
}
\examples{
library(cli)
letter <- style_italic(LETTERS[1:4])
letter2 <- col_blue(paste0(LETTERS[11:14],LETTERS[11:14],LETTERS[11:14]))
number <- 1:4
number2 <- col_red(10001:10004)
color <- c(col_blue("blue"),col_green("green"),col_red("red"),col_yellow("yellow"))
text <- c("The birch canoe slid on the smooth planks. Rice is often served in round bowls.",
          "Glue the sheet to the dark blue background. The juice of lemons makes fine punch.",
          "It's easy to tell the depth of a well. The box was thrown beside the parked truck.",
          "These days a chicken leg is a rare dish. The hogs were fed chopped corn and garbage.")
data <- data.frame(letter,letter2,number,number2,color,text)
cli_table(data)
}
