% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{coalesce}
\alias{coalesce}
\title{Coalesce}
\usage{
coalesce(...)
}
\arguments{
\item{...}{vectors, all the same size, listed in order of priority}
}
\value{
A single vector of values
}
\description{
Similar to SQL, use vectors in order until not NA.
}
\examples{
a <- c(1,  2,  NA, 4, NA)
b <- c(NA, NA, NA, 5, 6)
c <- c(7,  8,  NA, 9, 10)
coalesce(a,b,c)
}

