% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{coalesce_indicators}
\alias{coalesce_indicators}
\title{Coalesce indicator variables into one summary variable.}
\usage{
coalesce_indicators(..., else.value = NA_character_, first.hit.only = FALSE)
}
\arguments{
\item{...}{indicator variables to coalesce together. Should be all logical.}

\item{else.value}{The character value if there are no hits. Default is \code{NA}}

\item{first.hit.only}{If \code{TRUE}, will only show first hit (which is a true coalesce). Default is \code{FALSE}, which concatenates all hits.}
}
\value{
A vector of same length as the indicators, displaying variable names that were \code{TRUE}
}
\description{
After providing multiple indicator variables, summarize them by creating a character vector.
}
\examples{
#####
}
\author{
Ying Taur
}
