% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{coalesce_values}
\alias{coalesce_values}
\title{Coalesce values into one summary variable.}
\usage{
coalesce_values(...)
}
\arguments{
\item{...}{variables to coalesce together.}
}
\value{
A vector of same length as the variables, displaying variable names plus values.
}
\description{
After providing multiple indicator variables, summarize them by creating a character vector.
}
\examples{
#####
}
\author{
Ying Taur
}
