% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{color2hex}
\alias{color2hex}
\title{Convert color names to hex format}
\usage{
color2hex(color)
}
\arguments{
\item{color}{a vector of color names}
}
\value{
the color vector, all in hex format
}
\description{
Convert color names to hex format
}
\examples{
colors <- c("red","blue","yellow","purple")
color2hex(colors)
}
