% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{compare}
\alias{compare}
\title{Compare data frames}
\usage{
compare(x, y, by = NULL)
}
\arguments{
\item{x}{table to be compared}

\item{y}{table to be compared}

\item{by}{varaible(s) to join by and compare}
}
\value{
a table from full_join(x,y,by). Contains columns .status and .diffs.
}
\description{
Compare data frames
}
