% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{compare_relationship}
\alias{compare_relationship}
\title{Test the relationship between X and Y}
\usage{
compare_relationship(x, y)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}
}
\description{
Test the relationship between X and Y
}
