% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{compareOLD.info}
\alias{compareOLD.info}
\title{Compare info}
\usage{
compareOLD.info(x, ...)
}
\arguments{
\item{x}{first object}

\item{y}{second object}
}
\value{
list of values
}
\description{
Internal function for \code{compare()}.
}
