% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{convert.dates}
\alias{convert.dates}
\title{Search and Convert Date variables}
\usage{
convert.dates(data)
}
\arguments{
\item{data}{The data frame to be converted.}
}
\description{
In a given data frame, look for variables resembling dates and convert them to Dates.
}
\details{
Basically applies \code{as.Date2} to all variables.
}

