% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{copy.as.sql}
\alias{copy.as.sql}
\title{Convert vector to SQL code.}
\usage{
copy.as.sql(
  x,
  copy.clipboard = TRUE,
  fit = TRUE,
  width = getOption("width") - 15
)
}
\arguments{
\item{x}{vector to be converted to SQL code.}

\item{copy.clipboard}{logical, if \code{TRUE}, will copy the SQL code to the Clipboard.}
}
\value{
Returns the SQL code.
}
\description{
Produces SQL code for a vector of values.
}
\examples{
values <- c("35171234","35507574")
copy.as.sql(values)
}
\author{
Ying Taur
}
