% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{copy.as.tribble}
\alias{copy.as.tribble}
\title{Copy to clipboard as tribble}
\usage{
copy.as.tribble(tbl, spaces = 1)
}
\arguments{
\item{tbl}{a data frame to be copied}

\item{spaces}{number of spaces between columns}
}
\description{
Copy to clipboard as tribble
}
