% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{cummax}
\alias{cummax}
\alias{cummax.default}
\alias{cummax.Date}
\title{Cumulative Max}
\usage{
cummax(x, ...)

\method{cummax}{default}()

\method{cummax}{Date}(x)
}
\arguments{
\item{x}{vector, numeric or Date.}
}
\value{
Cumulative max of the vector
}
\description{
Overrides base::cummax in order to correctly manage dates, if that is the input.
}
\author{
Ying Taur
}
