% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{cut2}
\alias{cut2}
\title{Ying's Cut 2}
\usage{
cut2(x, lower, upper, n.quantiles, percentiles, lvls, rename.lvls = TRUE)
}
\arguments{
\item{x}{a numeric vector (or Date) to be converted to factor by cutting.}

\item{lower}{a vector of lower bounds}

\item{upper}{a vector of upper bounds}

\item{n.quantiles}{an integer specifying number of quantiles.}

\item{percentiles}{a vector of percentile breakpoints}

\item{lvls}{optional vector for renaming the levels}
}
\value{
a factor derived from grouping of x
}
\description{
Similar to cut, but with several options for grouping.
}
