% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{date.regex}
\alias{date.regex}
\title{Convert Date format to regular expression}
\usage{
date.regex(format)
}
\arguments{
\item{format}{character designating date formatting, following \code{\link[base:strptime]{base::strptime()}} convention.}
}
\value{
regular expression corresponding to the format
}
\description{
Convert Date format to regular expression
}
\examples{
date.regex("\%Y-\%m-\%d")
date.regex("\%a \%b \%e \%H:\%M:\%S \%Y")
}
