% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{deparse2}
\alias{deparse2}
\alias{copy.as.Rcode}
\title{Convert object to R-code.}
\usage{
deparse2(x, width = Inf)

copy.as.Rcode(x, width = getOption("width") - 15, copy.clipboard = TRUE)
}
\arguments{
\item{x}{object to be converted to R-code. Can be vector or data frame.}

\item{width}{max character width of each line. Set to \code{Inf} to avoid text-wrapping.}

\item{copy.clipboard}{whether or not to copy to clipboard. Default is \code{TRUE}}
}
\value{
Returns the R-code.
}
\description{
Produces R-code that would create the object inputted. I use this if I have some data object that I obtained
somehow but just want to declare it in the code.
}
\details{
This is similar to \code{deparse()}, except output looks a little bit more normal, and you can specify \code{width=Inf}
}
\examples{
x <- c("a","b","c")
copy.as.Rcode(x)
x <- tibble("a"=1:4,"_b"=c(T,F,T,F),"c c"=Sys.Date()+4:1,"d"=factor(LETTERS[1:4]))
deparse2(x)
}
\author{
Ying Taur
}
