% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{draw_key_polygon_close}
\alias{draw_key_polygon_close}
\title{draw_key_polygon_close}
\usage{
draw_key_polygon_close(data, params, size)
}
\description{
Same as \code{\link[ggplot2:draw_key]{ggplot2::draw_key_polygon()}}, but with zero gap width.
Used in \code{\link[=guide_gengrob.taxonomy]{guide_gengrob.taxonomy()}}.
}
