% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{dt}
\alias{dt}
\title{Ying's DT view}
\usage{
dt(
  data,
  fontsize = 14,
  pageLength = Inf,
  maxchars = 200,
  nrows = NULL,
  rownames = FALSE,
  escape = FALSE,
  whiteSpace = "pre-wrap"
)
}
\arguments{
\item{data}{dataframe to be viewed.}

\item{fontsize}{numeric controlling font size in the table, measured in px. Default is 11.}

\item{pageLength}{number of rows to display per page (Default \code{Inf}, show all rows)}

\item{maxchars}{max number of characters before adding an ellipsis \code{...}. Default is \code{200}.}

\item{nrows}{max number of rows to display. Default is \code{NULL},
which shrinks the data such that it falls under 1.5 Mb (the \code{\link[DT:datatable]{DT::datatable()}} limit).
Specify \code{Inf} to make sure there is no filtering.}

\item{rownames}{whether or not to show row names (passed directly to \code{\link[DT:datatable]{DT::datatable()}}). Default is \code{FALSE}.}

\item{escape}{whether to escape HTML entities in the table (passed directly to \code{\link[DT:datatable]{DT::datatable()}}).}

\item{whiteSpace}{CSS property sets how white space inside an element is handled.
Default is \code{"pre-wrap"}, which wraps text, and preserves whitespace and linebreaks.
Other options include: \code{"normal"}, \code{"nowrap"}, \code{"pre"}, \code{"pre-line"}}
}
\value{
A javascript-style datatable, which displays in the Rstudio viewer.
}
\description{
Use to peruse a data frame within RStudio. Utilizes \code{DT} package.
}
\details{
If data frame is grouped (i.e. \code{group_by} in dplyr), the rows will be sorted and shaded by group.
}
\examples{
library(dplyr)
mtcars \%>\% dt()
mtcars \%>\% group_by(cyl) \%>\% dt()
}
\author{
Ying Taur
}
