% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{dt_subset_data_for_viewing}
\alias{dt_subset_data_for_viewing}
\title{Find subset of data, for \code{dt()} viewing}
\usage{
dt_subset_data_for_viewing(df, max.size = 1500000)
}
\arguments{
\item{df}{data frame to be resized}

\item{max.size}{target object size in bytes. Default is \code{1500000}}
}
\value{
The row index corresponding to a total object size less than \code{max.size}.
}
\description{
Find subset of data, for \code{dt()} viewing
}
