% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{fan.angle}
\alias{fan.angle}
\title{Modify Angle for Text on Circular ggtree}
\usage{
fan.angle(angle, hjust = FALSE, right = FALSE)
}
\arguments{
\item{angle}{numeric vector of angles taken from ggtree data}

\item{hjust}{logical indicating whether to output the corresponding hjust parameter.}

\item{right}{logical indicating whether label should be at right angle to the tip being labelled.}
}
\value{
new calculated angle or hjust aesthetic
}
\description{
Modify Angle for Text on Circular ggtree
}
