% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{fct_add_freq}
\alias{fct_add_freq}
\title{Add Count/Prop to Factor}
\usage{
fct_add_freq(f, prop = FALSE)
}
\arguments{
\item{f}{factor to be modified}

\item{prop}{whether to include proportion (default is \code{FALSE})}
}
\value{
Factor with modified levels
}
\description{
Adds count or proportion to the levels of the factor.
}
\examples{
fct_add_freq(iris$Species)
}
