% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{file_activity_status}
\alias{file_activity_status}
\title{File activity status}
\usage{
file_activity_status(paths, wait.time = 1)
}
\arguments{
\item{paths}{the filepath(s) to be read.}

\item{wait.time}{amount of time in seconds to wait between reads. Default is 1 second.}
}
\value{
table describing files and their activity
}
\description{
Given filepath(s), detect changes in file sizes to see if it is actively changing.
}
