% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{fill_in_blanks}
\alias{fill_in_blanks}
\title{Fill in Blanks}
\usage{
fill_in_blanks(vec, blank = "", include.na = TRUE)
}
\arguments{
\item{vec}{the vector to be ed in.}

\item{blank}{vector of values that denote a blank. By default, \code{""} is used.}

\item{include.na}{vector of values that denote a blank. By default, \code{""} is used.}
}
\value{
Returns \code{vec}, with blanks filled in.
}
\description{
For a given vector,  in blanks with previous value.
}
\examples{
fill_in_blanks(c("1",NA,"2","","3","","","4",NA,NA))
}
\author{
Ying Taur
}
