% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{filter.phyloseq}
\alias{filter.phyloseq}
\title{Filter phyloseq object}
\usage{
\method{filter}{phyloseq}(phy, ..., prune_unused_taxa = TRUE, prune_unused_samples = FALSE)
}
\arguments{
\item{phy}{phylseq object}

\item{...}{filter criteria}

\item{prune_unused_taxa}{whether or not to remove unused taxa. Default is \code{TRUE}}

\item{prune_unused_samples}{whether or not to remove unused samples. Default is \code{FALSE}}
}
\description{
Subset a phyloseq using \code{dplyr}-style manipulation. Can be operations to be done on \code{sample_data} or \code{tax_table}.
}
\examples{
cid.phy \%>\% filter(Sample_ID!="1037",
                   Genus!="Blautia")
}
