% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{find.all.distinct.vars}
\alias{find.all.distinct.vars}
\title{Find All Distinct Variables}
\usage{
find.all.distinct.vars(data, ...)
}
\arguments{
\item{data}{data frame to be analyzed}

\item{...}{grouping variables that define data units.}
}
\value{
prints whether variables matching the groups or not.
}
\description{
Find Distinct
}
