% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\docType{data}
\name{fungal.palette}
\alias{fungal.palette}
\title{Fungal Palette}
\format{
An object of class \code{list} of length 8.
}
\usage{
fungal.palette
}
\description{
The customary palette for Fungi.
}
\keyword{datasets}
