% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{geom_hilight}
\alias{geom_hilight}
\title{Hilight a set of ggtree tips.}
\usage{
geom_hilight(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{oligo.file}{oligo file to be read. If oligo.file is a directory, all oligo files (*.oligos) will be read in.}
}
\value{
Returns a data frame containing oligo information
}
\description{
geom_hilight(aes(isTip=isTip,var=Genus,value="Blautia"))
geom_hilight(aes(isTip=isTip,var=Genus,value="Blautia"),xend=2)
geom_hilight(aes(isTip=isTip,var=Genus,value="Blautia"),xadd=1.5)
}
\examples{
...examples.here....
}
\author{
Ying Taur
}
\keyword{...}
\keyword{keyword1}
\keyword{keyword2}
