% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{geom_kaplanmeier}
\alias{geom_kaplanmeier}
\title{Generate Kaplan-Meier curve in ggplot2}
\usage{
geom_kaplanmeier(
  yvar,
  xvar = NULL,
  data,
  starttime = "tstart",
  flip.y = FALSE,
  size = NULL,
  logrank = FALSE,
  logrank.pos = NULL,
  logrank.fontsize = 5
)
}
\arguments{
\item{yvar}{character, used to indicate the variable set to be used as survival endpoint of interest.
For example, if \code{yvar="var1"} is specified, then \code{data} should have \code{"var1"} will represent
whether the endpoint occurred (logical or 0-1), and \code{"var1_day"} will represent the time at which
the event occurred (or didn't occur).}

\item{xvar}{character, indicating the variable within \code{data} that will contain the groups by which curves will be generated.}

\item{data}{data frame containing the survival data.}

\item{starttime}{character, specifying the column within \code{data} that indicates the survival start time.
If there is no left censoring, then this would refer to a vector of 0's. Default is \code{"tstart"}}

\item{flip.y}{logical, indicating whether or not to flip the y-axis. \code{flip.y = FALSE} by default (curve is downwards).}

\item{size}{line thickness for the survival curve.}
}
\value{
A ggplot2 geom object with Kaplan-Meier curve.
}
\description{
Creates a Kaplan-Meier curve which can be used like a geom in ggplot2.
}
\details{
Use this to make Kaplan-Meier curves in ggplot2. Utilizes the \code{geom_step} function to draw.
\code{yingtools::stcox} function is used to generate data from a survival frame.
}
\examples{
library(ggplot2)
ggplot() + geom_kaplanmeier("dead","intensity",data=cid.patients)
}
\author{
Ying Taur
}
