% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{geom_logrank}
\alias{geom_logrank}
\title{Generate label for Log-rank test results in ggplot2}
\usage{
geom_logrank(yvar, xvar, data, starttime = "tstart", pos, logrank.fontsize = 5)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Adds the p-value for a log-rank test to a ggplot2 graph.
}
\author{
Ying Taur
}
