% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{geom_taxonomy_interactive}
\alias{geom_taxonomy_interactive}
\title{Interactive timelines}
\usage{
geom_taxonomy_interactive(...)
}
\arguments{
\item{...}{arguments passed to base function, plus any of the interactive parameters}
}
\description{
Interactive timelines
}
\examples{
pt.meds <- cid.meds \%>\% filter(Patient_ID=="157") \%>\%
  mutate(tooltip=coalesce_values(med.clean,med.class,startday,endday,route,collapse="\n"))
g <- ggplot(pt.meds) +
  geom_timeline_interactive(aes(xmin=startday,xmax=endday,label=med.clean,by=med.class,fill=med.class,
                                tooltip=tooltip),check_overlap = TRUE)
g \%>\% girafe(ggobj=.,height_svg = 3.5)
}
