% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{geom_timebars}
\alias{geom_timebars}
\alias{scale_x_timebars}
\title{Time bars}
\usage{
geom_timebars(..., xlim, days, div = 30)

scale_x_timebars(..., xlim, days, div = 30)
}
\arguments{
\item{...}{arguments passed to \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}} and/or \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}}

\item{xlim}{2-value vector specifying the x-axis limits.}

\item{days}{vector of values where}

\item{div}{size of the bars, specified as reciprocal fraction of plot width. Default is 30 (1 bar is 1/30th of plot width).}
}
\description{
This geom and scale are used to transform the x-scale such that the scale moves slowly at certain points (\code{day}).
The x-scale can be numeric or date.
}
\examples{
data <- tibble(days=c(1,3,10,50))
xlim <- c(0,100)

ggplot(data) +
  geom_timebars(aes(x=days,y=1),xlim=xlim,days=data$days) +
  scale_x_timebars(xlim=xlim,days=data$days)
}
