% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\docType{data}
\name{GeomTaxonomy}
\alias{GeomTaxonomy}
\title{GeomTaxonomy}
\format{
An object of class \code{GeomTaxonomy} (inherits from \code{GeomCol}, \code{GeomBar}, \code{GeomRect}, \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTaxonomy
}
\description{
The ggproto object for stacked taxonomy.
}
\keyword{datasets}
