% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{get.abundance}
\alias{get.abundance}
\alias{add.abundance}
\title{Calculate Abundance from Phyloseq}
\usage{
get.abundance(phy, ..., denom = NULL)

add.abundance(phy, ..., denom = NULL)
}
\arguments{
\item{phy}{\code{\link[phyloseq:phyloseq-class]{phyloseq}} object to be analyzed}

\item{...}{one or more named taxonomic expressions defining the abundances to be calculated.
The expression should use variables from  \code{tax_table()} and evaluate to logical, to indicate which taxa are to be counted.}

\item{denom}{an expression specifying the denominator.
If there is no denominator, use \code{denom=NULL} (the default)
To calculate relative abundance, use \code{denom=TRUE}.}
}
\value{
a data frame containing sample identifier and abundance columns
}
\description{
Given a \code{\link[phyloseq:phyloseq-class]{phyloseq}}, calculate relative abundance for each sample.
\code{get.abundance} returns data frame with values, \code{add.abundance} returns phyloseq with values stored in \code{sample_data}
}
\examples{
library(phyloseq)
get.abundance(cid.phy,
              entero.nseqs=Genus=="Enterococcus",
              proteo.nseqs=Phylum=="Proteobacteria",
              denom=TRUE)
get.abundance(cid.phy,
              entero.pct=Genus=="Enterococcus",
              proteo.pct=Phylum=="Proteobacteria",
              denom=TRUE)
get.abundance(cid.phy,
              bact.firm.ratio=Phylum=="Bacteroidetes",
              denom=Phylum=="Firmicutes")
}
