% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{get.all.dependencies}
\alias{get.all.dependencies}
\title{Determine all dependent packages}
\usage{
get.all.dependencies(pkg)
}
\arguments{
\item{pkg}{package name}
}
\value{
a character vector of all sub-packages
}
\description{
For a given installed package, determine all the downstream dependent packages, nesting through all sub-packages.
Removes R itself and its included base packages, such as \code{stats} or \code{utils}.
}
\details{
This function works by searching through dependencies in a recursive manner. A global data frame is maintained to remove packages that
have already been processed. This increases the efficiency and avoids any possibility of circular dependencies (if that is even possible).
}
\examples{
get.all.dependencies("purrr")
}
