% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{get.code.info}
\alias{get.code.info}
\title{Get Code Info}
\usage{
get.code.info(expr, text = NULL, fn = NULL, recursive = FALSE)
}
\arguments{
\item{expr}{expression to be evaluated.}

\item{text}{character, code can alternatively input as a string.}

\item{fn}{function, use this to evaluate code within a function.}

\item{recursive}{whether to examine code within functions}
}
\value{
list containing information such as packages, functions, parsedata.
}
\description{
Read code (as an expression, text, or function) and provide information on
functions, package dependencies
}
\examples{
get.code.info(overlaps(1,2,3,sqrt(44)))

get.code.info(text="log_epsilon_trans(0.001)")
yingtools2::age.years()

fun <- function() {
  age.years(as.Date("1975-02-21"),Sys.Date())
}
get.code.info(fn=fun)
}
