% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{get.dist}
\alias{get.dist}
\title{Convert pairwise distance table to distance matrix}
\usage{
get.dist(pw)
}
\arguments{
\item{pw}{a data frame of pairwise distances.}
}
\value{
corresponding distance matrix
}
\description{
This is the inverse function of {get.pairwise()}
}
\examples{
pw <- get.pairwise(dist(mtcars))
get.dist(pw)
}
