% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{get.distance}
\alias{get.distance}
\title{Calculate distance matrix from phyloseq data}
\usage{
get.distance(phy, method, ...)
}
\arguments{
\item{phy}{phyloseq object}

\item{method}{character string indicating distance metric to be calculated. Can be a method from
\code{\link[phyloseq:distanceMethodList]{phyloseq}}, or \verb{"taxhorn'}}

\item{...}{passed to \code{\link[phyloseq:distance]{phyloseq::distance()}}}
}
\value{
a distance metric
}
\description{
Basically same as \code{\link[phyloseq:distance]{phyloseq::distance()}}, but adds \code{taxhorn} metric
}
