% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{get.environment.locations}
\alias{get.environment.locations}
\title{Get possible R environment path locations.}
\usage{
get.environment.locations()
}
\description{
Get possible R environment path locations.
}
