% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{get.otu.melt}
\alias{get.otu.melt}
\title{Convert Phyloseq to Melted OTU x Sample Data}
\usage{
get.otu.melt(phy, filter.zero = TRUE, sample_data = TRUE)
}
\arguments{
\item{phy}{phyloseq object containing sample data}

\item{filter.zero}{Logical, whether or not to remove zero abundances. Default \code{TRUE}.}

\item{sample_data}{Logical, whether or not to join with \code{sample_data}. Default \code{TRUE}.}
}
\value{
Data frame melted OTU data
}
\description{
Creates OTU+Sample-level data, using phyloseq object (ID=otu+sample)
}
\details{
Essentially gives back the OTU table, in melted form, such that each row represents a certain OTU for a certain sample.
Adds sample and taxonomy table data as columns. Uses the following reserved varnames: otu, sample, numseqs, pctseqs.
Note that phyloseq has a similar function, \code{psmelt}, but that takes longer.
The \code{get.otu.melt} now works by performing operations via data table, making it about 30x faster than before.
}
\examples{
library(phyloseq)
get.otu.melt(cid.phy)
}
