% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{get.pairwise}
\alias{get.pairwise}
\title{Convert Distance Matrix to Pairwise Distances}
\usage{
get.pairwise(dist)
}
\arguments{
\item{dist}{distance matrix to be converted}

\item{diag}{whether or not to include diagonal}
}
\value{
data frame comtaining pairwise distances
}
\description{
This is the inverse function of \code{\link[=get.dist]{get.dist()}}
}
\examples{
get.pairwise(dist(mtcars))
}
