% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{get.phyloseq.from.melt}
\alias{get.phyloseq.from.melt}
\title{Convert melted OTU table to phyloseq object}
\usage{
get.phyloseq.from.melt(
  otu.melt,
  sample_id = "sample",
  abundance_var = "numseqs",
  taxa_id = "otu",
  taxranks = c("Superkingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  sample_vars = NULL
)
}
\arguments{
\item{otu.melt}{table of taxa x sample abundances, similar to output of \code{get.otu.melt}}

\item{sample_id}{sample ID variable. Default \code{"sample"}.}

\item{abundance_var}{abundance variable, used to fill \code{otu_table()}. Default \code{"sample"}.}

\item{taxa_id}{taxa/OTU ID variable. Default \code{"otu"}.}

\item{taxranks}{vector of taxonomic ranks to be included in \code{tax_table()}.}

\item{sample_vars}{vector of sample variables to be included in \code{sample_data()}. If \code{NULL}, will attempt to determine sample vars automatically (vars that are distinct along with \code{sample_id}).}
}
\value{
phyloseq object, generated from the \code{otu.melt} data.
}
\description{
Convert melted OTU table to phyloseq object
}
\examples{
library(phyloseq)
phy <- cid.phy
ranks <- rank_names(phy)
otu <- get.otu.melt(cid.phy)
phy2 <- get.phyloseq.from.melt(otu,taxranks=ranks)
phy
phy2
}
