% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{get.samp}
\alias{get.samp}
\title{Extract Phyloseq sample_data}
\usage{
get.samp(phy, stats = FALSE, measures = c("Observed", "InvSimpson",
  "Shannon"))
}
\arguments{
\item{phy}{phyloseq object containing sample_data}

\item{stats}{logical, whether or not to include summary statistics of samples. Stores nseqs, and diversity metrics.}

\item{measures}{diversity measures to calculate, if stats is TRUE. Default: c("Observed","InvSimpson","Shannon")}
}
\value{
Dataframe containing sample_data data
}
\description{
Creates data.frame from sample_data, storing the rownames as variable "sample. The opposite of function set.samp.
}

