% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{get.tax.legend}
\alias{get.tax.legend}
\title{Generate tax legend}
\usage{
get.tax.legend(tax.palette = yt.palette3, fontsize = 5)
}
\arguments{
\item{tax.palette}{a list of formulas specifying the palette. Default is \code{\link{yt.palette3}}.}

\item{fontsize}{Font size. Default is 5.}
}
\value{
a ggplot object showing the legend.
}
\description{
Generate tax legend
}
