% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get.tax.legend2}
\alias{get.tax.legend2}
\title{Generate tax legend 2}
\usage{
get.tax.legend2(
  tax.palette = yt.palette3,
  fontsize = 5,
  ncol = NULL,
  nrow = NULL
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
