% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{get.tax}
\alias{get.tax}
\title{Extract Phyloseq tax_table}
\usage{
get.tax(phy)
}
\arguments{
\item{phy}{phyloseq object containing tax_data}
}
\value{
Dataframe containing tax data
}
\description{
Creates data.frame from tax_table, storing the rownames as variable "otu". The opposite of set.tax function.
}
