% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{get.time}
\alias{get.time}
\title{Extract Time}
\usage{
get.time(datetime, format = "\%I:\%m\%p")
}
\arguments{
\item{datetime}{POSIXct object}

\item{format}{character parameter for formatting of time. Default format "10:30AM"}
}
\value{
character with time component
}
\description{
Get time from a date-time POSIXct object.
}
