% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{get.yt.palette2}
\alias{get.yt.palette2}
\title{Get YT Palette 2}
\usage{
get.yt.palette2(tax)
}
\arguments{
\item{tax}{either a data.frame, phyloseq, or tax_table}

\item{use.cid.colors}{whether to use classic CID colors}
}
\value{
a color palette that can be used in \code{ggplot2}
}
\description{
Get YT Palette 2
}
\author{
Ying Taur
}
