% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{get_epsilon}
\alias{get_epsilon}
\title{Get a reasonable epsilon value for log_epsilon_trans()}
\usage{
get_epsilon(x, prob = 0.2)
}
\arguments{
\item{x}{vector of values}

\item{prob}{proportion of values in \code{x} that will fall below epilon}
}
\value{
epsilon value
}
\description{
Get a reasonable epsilon value for log_epsilon_trans()
}
