% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_release.R
\name{get_gitdir_info}
\alias{get_gitdir_info}
\title{Get git info on all git folders}
\usage{
get_gitdir_info(parent.dir = ".")
}
\arguments{
\item{parent.dir}{path containing all github folders}
}
\value{
Table containing data on all git folders
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
get_getdir_info("C:/Users/Ying/R")
}
