% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{get_taxonomy_info}
\alias{get_taxonomy_info}
\title{Get Information on Taxonomy Lineage}
\usage{
get_taxonomy_info(phy, pct.cutoff = 0.99)
}
\arguments{
\item{pct.cutoff}{}
}
\description{
For each taxonomy level, calculate statistics across samples. Also provides grouping
variable, for levels that are essentially equivalent.
}
