% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{gg.axis.limits}
\alias{gg.axis.limits}
\title{Calculate axis limits}
\usage{
gg.axis.limits(gg)
}
\arguments{
\item{gg}{the ggplot object}
}
\value{
a list containing inforation about limits for X and Y.
}
\description{
Determines the actual limits of X and Y, for a given ggplot object. This is used by \code{\link[=gg.align.xlim]{gg.align.xlim()}}.
}
