% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{gg.colors}
\alias{gg.colors}
\title{Default color palette of ggplot2}
\usage{
gg.colors(n = 6, h = c(0, 360) + 15)
}
\arguments{
\item{n}{number of colors to display}

\item{h}{The hue of the color specified as an angle in the range [0,360]. 0 yields red, 120 yields green 240 yields blue, etc.}
}
\value{
A color palette that would have been used by ggplot2 by default
}
\description{
The color palette that ggplot2 uses by default.
}
\examples{
gg.colors(6)
}
\author{
Ying Taur
}

