% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{gg.stack2}
\alias{gg.stack2}
\title{Stack and line up ggplot objects in a column}
\usage{
gg.stack2(
  ...,
  heights = NULL,
  gap = unit(1, "pt"),
  margin = NULL,
  adjust.themes = TRUE,
  return.gg.list = FALSE
)
}
\arguments{
\item{...}{ggplot objects to be stacked. Can also supply a formula where left hand side is ggplot, right hand side is height.}

\item{heights}{a numeric vector representing the relative height of each plot. Passed directly to \code{\link[gridExtra:arrangeGrob]{gridExtra::grid.arrange()}}.}

\item{gap}{size of gap between stacked plots. Default is \code{unit(0,"pt")}.}

\item{margin}{size of the margin around the plots. Default is \code{unit(5.5,"pt")}, which is the \code{ggplot2} default.}

\item{adjust.themes}{logical, whether or not to adjust each plot's theme for stacking (change gap/margin, suppress x-axis in upper plots). Default \code{TRUE}.}

\item{gg.extras}{a list of ggplot objects that will be applied to all plots. Default is \code{NULL}.}
}
\value{
plot of stacked ggplots
}
\description{
Use this to arrange ggplot objects, where the axes, plot, and legend are lined up correctly.
}
\details{
Makes use of \code{\link[=patchwork-package]{patchwork}} package to align. The previous version (\code{\link[=gg.stack]{gg.stack()}})
manually adjusted widths to perform the aligment.
}
\examples{
g1 <- ggplot(mtcars,aes(x=mpg,y=disp,color=factor(cyl))) + geom_point()
g2 <- ggplot(mtcars,aes(x=mpg,y=wt,fill=factor(cyl))) + geom_col() + scale_fill_discrete("Number of Cylinders")
g3 <- ggplot(mtcars,aes(x=mpg,y=wt,label=cyl,fill=factor(cyl))) + geom_label()

# grid.arrange does not align correctly, basically due to because of legend/axis differences
gridExtra::grid.arrange(g1,g2,g3,ncol=1)
# gg.stack2 aligns correctly
gg.stack2(g1,g2,g3)
# vary the heights
gg.stack2(g1,g2,g3,heights=c(1,2,3))
# alternatively, use formulas to specify
gg.stack2(g1~3,
         g2~2,
         g3~1,heights=c(1,2,3))
}
