% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{\%-\%}
\alias{\%-\%}
\title{Subtract Dates}
\usage{
x \%-\% y
}
\arguments{
\item{x}{vector to subtract from}

\item{y}{vector to subtract}
}
\value{
vector representing `x - y`
}
\description{
Returns number of days.
}
\examples{
Sys.Date() \%-\% as.Date("1975-02-21")
}
