% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{group_by_all_distinct}
\alias{group_by_all_distinct}
\title{Group By All Distinct Variables}
\usage{
group_by_all_distinct(data, ...)
}
\arguments{
\item{data}{data frame}

\item{...}{variables to group by}
}
\value{
Returns \code{data}, but grouped by \code{...} plus other variables that can be grouped along with it.
}
\description{
Can be used similar to \code{group_by}, but will try to add additional variables to the group list, such that the grouping remains the same.
In other words, \code{group_by_all_distinct(data,a,b,c)} will group by a,b,c,x,y,z, where x,y,z do not alter the groups.
This is useful for keeping extra variables that go with the grouping, if you perform \code{summarize} afterwards.
}
\details{
This is a convenience function that I made because of sheer laziness....
probably better to avoid using this for really rigorous data operations.
}
\author{
Ying Taur
}
