% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{group_split_named}
\alias{group_split_named}
\title{Split data frame into named list}
\usage{
group_split_named(.tbl, ..., .keep = TRUE, sep = " / ")
}
\arguments{
\item{.tbl}{A table}

\item{...}{Grouping specification}

\item{.keep}{Should the grouping columns be kept}

\item{sep}{If multiple grouping vars are specified, separate the values with this character separator.}
}
\value{
A list of tibbles, broken up by grouping.
}
\description{
Same as \code{dplyr::group_split()}, except that the list of tables is named rather than unnamed.
Another minor difference is that it keeps previous grouping (\code{group_split()} wants all grouping vars to be stated at once).
}
\examples{
group_split_named(iris,Species)
}
