% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{guide_gengrob.taxonomy}
\alias{guide_gengrob.taxonomy}
\title{guide_gengrob.taxonomy (OLD)}
\usage{
\method{guide_gengrob}{taxonomy}(guide, theme)
}
\description{
Called by name, by \code{\link[=guide_taxonomy]{guide_taxonomy()}}.
}
