% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{guide_taxonomyOLD}
\alias{guide_taxonomyOLD}
\title{Taxonomy Guide (OLD)}
\usage{
guide_taxonomyOLD(
  title = waiver(),
  override.tax.palette = NULL,
  title.position = NULL,
  title.theme = NULL,
  title.hjust = NULL,
  title.vjust = NULL,
  label = TRUE,
  label.position = NULL,
  label.theme = NULL,
  label.hjust = NULL,
  label.vjust = NULL,
  keywidth = 3,
  keyheight = NULL,
  direction = NULL,
  default.unit = "line",
  override.aes = list(),
  nrow = NULL,
  ncol = NULL,
  byrow = FALSE,
  reverse = FALSE,
  order = 0,
  ...
)
}
\description{
Taxonomy Guide (OLD)
}
