% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{guide_train.taxonomy}
\alias{guide_train.taxonomy}
\title{guide_train.taxonomy}
\usage{
\method{guide_train}{taxonomy}(guide, scale, aesthetic = NULL)
}
\description{
Adopted from \code{ggplot2:::guide_train.legend}, but converts items to list-col colors.
Called by name, by \code{\link[=guide_taxonomy]{guide_taxonomy()}}.
}
