% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{highlight_grep.OLD}
\alias{highlight_grep.OLD}
\title{Highlight Text Using Regex}
\usage{
highlight_grep.OLD(
  text,
  ...,
  ignore_case = TRUE,
  color_fun = ~str_c("<b><font color=\\"", .y, "\\">", .x, "</font></b>"),
  gap_nchars = Inf,
  return_df = FALSE
)
}
\arguments{
\item{text}{the character vector to be searched.}

\item{...}{regular expression patterns to be searched.
Can also supply list(matrix) of locations (output from \code{\link[stringr:str_locate]{stringr::str_locate_all()}})
Arguments can be named, in order to specify color.
If unnamed, default colors are supplied.}

\item{ignore_case}{Whether or not to ignore case. Default is \code{TRUE}}

\item{color_fun}{Function (purrr-style) that modifies pattern hits. Takes text and color as arguments.
Default is \verb{~str_c("<b><font color=\\"",.y,"\\">",.x,"</font></b>")}.}

\item{gap_nchars}{Optional integer, truncate text around regex hits, by specified number of characters.
Default is \code{Inf} (no truncation).}

\item{return_df}{Optional, if \code{TRUE}, will return the data frame, rather than the replaced text.}
}
\value{
Modified form of \code{text}, with all regex hits replaced.
}
\description{
Search a text vector and highlight the hits with HTML color.
}
\details{
Note that overlapping regex hits will be colored by mixing the assigned colors together.
}
\examples{
highlight_grep(sentences[1:3],"red"="the","blue"="background|planks","green"="depth|back|sheet")
}
