% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{highlight_text}
\alias{highlight_text}
\alias{highlight_html}
\title{Find and Highlight Text}
\usage{
highlight_text(
  text,
  pattern,
  before = "###",
  after = "###",
  ignore_case = TRUE
)

highlight_html(text, pattern, color = "red", ignore_case = TRUE)
}
\arguments{
\item{text}{character vector of text to search and modify}

\item{pattern}{regular expression pattern to search}

\item{before}{emphasizing text to add before search hits.}

\item{after}{emphasizing text to add after search hits.}

\item{ignore_case}{Should case differences be ignored? Default is \code{TRUE}.}
}
\value{
a modified form of \code{text} is returned.
}
\description{
Search text using regular expression pattern and add something around it to emphasize.
}
\examples{
highlight_text(sentences,"the")

d <- tibble(sentences) \%>\%
  mutate(sentences=highlight_html(sentences,"the"))
dt(d)

\dontrun{
d <- tibble(sentences) \%>\%
  mutate(sentences=highlight_html(sentences,"the"))
dt(d)
}
}
