% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\alias{info,ANY-method}
\alias{info,data.frame-method}
\alias{compare}
\alias{compare,ANY,ANY-method}
\title{Info on an object, or compare two objects}
\usage{
info(x, ...)

\S4method{info}{ANY}(x)

\S4method{info}{data.frame}(x, by = NULL)

compare(x, y, ...)

\S4method{compare}{ANY,ANY}(x, y)
}
\arguments{
\item{x}{object to be described/compared.}

\item{y}{object to be compared.}
}
\description{
Provide information about an object, or compare two objects.
}
\details{
\if{html}{\out{<div class="sourceCode asciicast">}}\preformatted{x <- c("A","B","C")
cli_alert("letters are \{x\}")
}\if{html}{\out{</div>}}
}
\examples{
library(tidyverse)
chr <- sentences[1:10]
chr.rev <- rev(chr)
chr.dup <- c(chr,chr,chr)
chr.dup2 <- c(chr,chr) \%>\% sort()
chr.part <- chr[1:5]
num <- 1:10
dates <- Sys.Date() + 1:10
fct <- as.factor(chr)
fct.rev <- fct_rev(fct)
lgl <- c(T,T,F,F,T,T,T,F,F,T)
lst <- as.list(df)
mat <- as.matrix(mtcars)
df <- mtcars \%>\% rownames_to_column("car")
df.dup <- bind_rows(df,df)
df.sort <- df \%>\% arrange(mpg)
df.partrows <- df \%>\% slice(1:6)
df.partcols <- df \%>\% select(mpg,cyl,disp) \%>\% mutate(col=1)
df.part1 <- df \%>\% select(car,2:9) \%>\% slice(1:22)
df.part2 <- df \%>\% select(car,5:12) \%>\% slice(15:32)
df.diff <- df \%>\% mutate(mpg=ifelse(row_number()==1,1,mpg))
df.diff2 <- df \%>\% mutate(mpg=mpg+1,disp=disp+1)
df.nojoin <- df \%>\% mutate(car=paste(car,"x"))
info(chr)
info(chr.dup)
info(chr.rev)
info(chr.part)
info(num)
info(dates)
info(fct)
info(lgl)
info(df)
info(df,by="car")
info(df,by=c("mpg","car"))
compare(chr,chr)
compare(chr,chr.rev)
compare(chr,chr.dup)
compare(chr.dup,chr.dup2)
compare(fct,fct.rev)
compare(df,df)
compare(df,df.sort)
compare(df,df.partcols)
compare(df,df.partrows)
compare(df.partcols,df.partrows)
compare(df,df.partrows,by="car")
compare(df,df.nojoin,by="car")
compare(df,df.dup,by="car")
compare(df.part1,df.part2,by="car")
compare(df,df.diff,by="car")
compare(df,df.diff2,by="car")
}
