% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{inline}
\alias{inline}
\title{Modify Shiny Input Element to Inline Label}
\usage{
inline(..., padding = "5px")
}
\arguments{
\item{element}{expression for shiny element to be modified.}
}
\description{
Change a Shiny element such that label is printed horizontally, instead of vertically.
}
\examples{
inline(textInput("text1","text input"),
       numericInput("number1","number input",value=1))
}
