% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{inner_join_fuzzy}
\alias{inner_join_fuzzy}
\title{Fuzzy Inner Join}
\usage{
inner_join_fuzzy(
  x,
  y,
  by = NULL,
  suffix = c(".x", ".y"),
  ignore.case = TRUE,
  max.distance = 1
)
}
\arguments{
\item{x}{first data frame to be joined}

\item{y}{second data frame to be joined}

\item{by}{the columns to be joined by.}

\item{suffix}{suffixes to be used if there are duplicate names between \code{x} and \code{y}.}

\item{ignore.case}{whether to ignore case (default is \code{TRUE})}

\item{max.distance}{max Levenshtein distance allowed (default is \code{1})}
}
\value{
A joined data frame from \code{x} and \code{y}. \code{dist_} will be stored, representing the Levenshtein distance.
}
\description{
Inner join two data frames using fuzzy matching via \code{\link[=adist]{adist()}}.
}
\examples{
x <- tibble(name=c("Luke Skywalker", "Darth Vader", "Greedo"),
            species=c("Human", "Human", "Rodian"))
y <- tibble(name=c("Luke Skywalker", "Darth Vaderx", "Chewbacca"),
            homeworld=c("Tatooine", "Tatooine", "Kashyyyk"))
inner_join_fuzzy(x,y,by="name")
inner_join_fuzzy(x,y,by="name",max.distance = 10)
}
