% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{inner_join_replace_old}
\alias{inner_join_replace_old}
\alias{left_join_replace_old}
\alias{right_join_replace_old}
\alias{full_join_replace_old}
\title{Inner/Left/Right/Full Join with Replace}
\usage{
inner_join_replace_old(
  x,
  y,
  by = NULL,
  conflict = c("yonly", "xonly", "ycoalesce", "xcoalesce", "error")
)

left_join_replace_old(
  x,
  y,
  by = NULL,
  conflict = c("yonly", "xonly", "ycoalesce", "xcoalesce", "error")
)

right_join_replace_old(
  x,
  y,
  by = NULL,
  conflict = c("yonly", "xonly", "ycoalesce", "xcoalesce", "error")
)

full_join_replace_old(
  x,
  y,
  by = NULL,
  conflict = c("yonly", "xonly", "ycoalesce", "xcoalesce", "error")
)
}
\arguments{
\item{x}{first data frame to be joined}

\item{y}{second data frame to be joined}

\item{by}{a character vector of variables to be joined by.}

\item{conflict}{what to do if columns conflict.
\enumerate{
\item \code{y} always keep the y-value (default).
\item \code{x} always keep the x-value.
\item \code{y.coalesce} keep the y-value unless it is \code{NA}.
\item \code{x.coalesce} keep the x-value unless it is \code{NA}.
\item \code{error} throw error if there is a conflict.
}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Same as \code{inner_join}, \code{left_join}, \code{right_join}, and \code{full_join} in the \code{dplyr} package, except that variables with the
same column name will not be renamed with the ".x" and ".y" suffix.
Instead, the variables will be turned into one column if the variables are equal. If they are not equal, an error (or warning) is thrown.
}
\details{
This is a convenience function that just avoids the renaming of columns.
}
\examples{
tbl1 <- tibble(id=1:10) \%>\% mutate(source="table1")
tbl2 <- tibble(id=5:15) \%>\% mutate(source="table2")
full_join(tbl1,tbl2,by="id")
full_join_replace(tbl1,tbl2,by="id") \%>\% arrange(id)
}
