% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{ir_dist}
\alias{ir_dist}
\alias{irl_dist}
\title{Find Distance Between Two \code{IRanges}}
\usage{
ir_dist(ir1, ir2)

irl_dist(irl1, irl2)
}
\arguments{
\item{ir1}{First \code{IRanges} object}

\item{ir2}{Second \code{IRanges} object}

\item{irl1}{First \code{IRangesList} object}

\item{irl2}{Second \code{IRangesList} object}
}
\value{
An integer vector of distances.
Positive distance means \code{ir1} came before \code{ir2},
Negative distance means \code{ir2} came before \code{ir1},
Zero distance means the ranges overlap.
}
\description{
Calculate closest distance between two \code{IRanges} or \code{IRangesList} objects.
}
