% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{is.between}
\alias{is.between}
\title{Determines if x is between start and stop/}
\usage{
is.between(x, start, stop)
}
\description{
Similar to \code{dplyr::between}, except that the vectors are recycled, so x can be a fixed value.
}
