% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{is.distinct}
\alias{is.distinct}
\title{Is Distinct}
\usage{
is.distinct(data, ..., add.group.vars = TRUE)
}
\arguments{
\item{data}{Dataframe to be analyzed}

\item{...}{grouping variables that define data units.}

\item{add.group.vars}{if \code{TRUE}, add any grouping variables.}
}
\value{
Logical indicating whether or not columns are distinct.
}
\description{
Determine if specified columns within data are distinct for individual rows.
}
\examples{
cid.patients \%>\% is.distinct(Patient_ID)
cid.cdiff \%>\% is.distinct(Patient_ID)
}
