% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{is.mrn}
\alias{is.mrn}
\title{Determine if variable is a properly formatted MSKCC MRN}
\usage{
is.mrn(mrn, like = FALSE)
}
\arguments{
\item{mrn}{vector to be examined.}

\item{like}{logical, whether or not to check if the variable \emph{could} be an MRN. If \code{FALSE}, it checks strictly. If \code{TRUE}, will allow for classes other than character, and for whitespaces.}
}
\value{
Returns logical stating whether or not this variable is an MRN.
}
\description{
MRN's must characters with 8 digits, where the first is either "0" or "3".
}
\details{
Note that many numeric vectors \emph{could} be an MRN. Will issue a warning if the variable
passes the tests but is not necessarily an MRN variable.
}
\author{
Ying Taur
}
