% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{is.one.to.one}
\alias{is.one.to.one}
\title{Test if X and Y are one-to-one}
\usage{
is.one.to.one(x, y)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}
}
\value{
logical value as to whether or not the vectors are one-to-one.
}
\description{
Test if X and Y are one-to-one
}
