% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{is.pattern}
\alias{is.pattern}
\title{Check if object is a Regex pattern}
\usage{
is.pattern(x)
}
\arguments{
\item{x}{Object to be inspected}
}
\description{
Checks if object is a single character vector, or is a \code{stringr}-style \code{\link[stringr:modifiers]{pattern}}
}
\examples{
re.letter <- "[a-z]+"
re.number <- "[0-9]+"
is.pattern(re.letter)
is.pattern(re.number)
is.pattern(regex(re.letter,ignore_case=TRUE))
is.pattern(123)
}
