% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Determines if a numeric is all whole numbers}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{numeric vector to be analyzed.}

\item{tol}{tolerance for whole number calling. Usually no need to change this.}
}
\value{
Returnes logical value of whether or not all values are whole numbers.
}
\description{
Uses machine precision to determine if a numeric vector is all whole numbers.
}
\examples{
a <- c(1,2,3)
b <- c(1,2,3.00001)
c <- c(1,2,3.00000000001)
is.wholenumber(a)
is.wholenumber(b)
is.wholenumber(c)
}
\author{
Ying Taur
}
