% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{kill_port_process}
\alias{kill_port_process}
\title{Kill port process}
\usage{
kill_port_process(port)
}
\arguments{
\item{port}{the port (numeric) to search.}
}
\description{
Find process that is listening to the specified port and kill it.
}
\examples{
kill_port_process(4567)
}
