% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\docType{data}
\name{LayerTaxonomy}
\alias{LayerTaxonomy}
\title{LayerTaxonomy}
\format{
An object of class \code{LayerTaxonomy} (inherits from \code{Layer}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
LayerTaxonomy
}
\description{
The Layer object for \code{\link[=geom_taxonomy]{geom_taxonomy()}}.
}
\details{
Two things are done here:
\enumerate{
\item If \code{scale_fill} is not specified, the \code{\link[=scale_fill_taxonomy]{scale_fill_taxonomy()}} scale is applied.
\item The fill aesthetic is converted to a factor, ordered by \code{tax.palette} colors,
to allow for stacking in the correct order.
}
}
\keyword{datasets}
