% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{lda.clado}
\alias{lda.clado}
\title{Plot LDA Results in Cladogram}
\usage{
lda.clado(lda, layout = "circular", pad = 2, check_overlap = TRUE)
}
\arguments{
\item{lda}{lda table from \code{lda.effect()}}

\item{layout}{Either "cirular" or "rectangular"}

\item{pad}{Determines the spacing of surrounding clade labels.}
}
\value{
ggplot of lda cladogram
}
\description{
Plot LDA Results in Cladogram
}
