% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{lda.effect}
\alias{lda.effect}
\title{LDA Effect Size}
\usage{
lda.effect(
  phy,
  class,
  subclass = NULL,
  subject = NULL,
  anova.alpha = 0.05,
  wilcoxon.alpha = 0.05,
  lda.cutoff = 2,
  wilcoxon.within.subclass = FALSE,
  one.against.one = FALSE,
  n_boots = 30,
  min_c = 10,
  f_boots = 0.67,
  mult.test.correction = 0
)
}
\arguments{
\item{phy}{the phyloseq object containing the data}

\item{class}{set which variable use as class}

\item{subclass}{set which variable use as subclass (default is no subclass)}

\item{anova.alpha}{set the alpha value for the Kruskal-Wallis test (default 0.05)}

\item{wilcoxon.alpha}{set the alpha value for the Wilcoxon test (default 0.05)}

\item{lda.cutoff}{set the threshold on the absolute value of the logarithmic LDA score (default 2.0)}

\item{wilcoxon.within.subclass}{set whether perform the wilcoxon test only among the subclasses with the same name (default FALSE)}

\item{one.against.one}{(for multiclass tasks) set whether the test is performed in a
one-against-one (TRUE - more strict) or in a one-against-all setting (FALSE - less strict) (default FALSE)}

\item{n_boots}{set the number of bootstrap iteration for LDA (default 30). Set to NULL to skip bootstrapping altogether.}

\item{min_c}{minimum number of samples per subclass for performing wilcoxon test (default 10)}

\item{f_boots}{set the subsampling fraction value for each bootstrap iteration (default 0.67)}

\item{mult.test.correction}{set the multiple testing correction options. 0 no correction (more strict, default),
1 correction for independent comparisons, 2 correction for dependent comparison}
}
\value{
a table containing features tested and results.
}
\description{
Given a phyloseq object and class variable, perform LDA Effect Size analysis.
}
\examples{
lda <- lda.effect(cid.phy,class="Consistency")
}
