% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{lda.plot}
\alias{lda.plot}
\title{Plot LDA Results}
\usage{
lda.plot(lda, tax.label = "taxon")
}
\arguments{
\item{lda}{lda table from \code{lda.effect()}}

\item{tax.label}{either "taxon" or "taxonomy"}
}
\value{
a ggplot of lda results
}
\description{
Plot LDA Results
}
