% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{left_join_replace}
\alias{left_join_replace}
\title{Left Join Replace}
\usage{
left_join_replace(x, y, by, suffix = c(".x", ".y"), ...)
}
\arguments{
\item{var}{the vector to be tabulated}

\item{sortby.freq}{if \code{TRUE}, sorts by order}

\item{useNA}{character specifying whether to tally \code{NA} values. This is passed to \code{tabulate}}

\item{as.char}{logical specifying whether to return tabulation as a single character. Useful for summarizing data within grouping commands such as \code{ddply} or \code{group_by}/\code{summarize}}
}
\value{
Returns a data frame with tabulations.
}
\description{
Like left_join, but replaces column names in x with new columns from y.
}
\examples{
...examples.here....
}
\author{
Ying Taur
}
