% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{list.all.exported.functions}
\alias{list.all.exported.functions}
\title{List all functions exported by installed packages}
\usage{
list.all.exported.functions(remove.reexported = TRUE)
}
\arguments{
\item{remove.reexported}{If \code{TRUE} (default), remove functions that were re-exported from other packages. (e.g. \code{ggtree::ggplot} is the same as \code{ggplot2::ggplot})}
}
\description{
Find all exported functions from all installed packages. Use this to find namespace collisions.
}
\examples{
f <- list.all.exported.functions()
f \%>\% filter(Function=="rename")
f \%>\% filter(Function=="slice")
f \%>\% filter(Function=="rescale")
f \%>\% filter(Function=="first")
}
