% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{list.gg.items}
\alias{list.gg.items}
\title{List items within ggplot object}
\usage{
list.gg.items(ggobj, prefix = NULL, depth = 1)
}
\arguments{
\item{ggobj}{ggplot object}

\item{prefix}{for internal use, to track super.}

\item{depth}{for internal use, to keep track of depth.}
}
\value{
a data frame listing of all objects.
}
\description{
A recursive function that lists all objects in a \code{ggplot} object.
}
