% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{locs2irl}
\alias{locs2irl}
\alias{irl2locs}
\alias{irl2df}
\alias{df2irl}
\alias{locs2df}
\alias{df2locs}
\alias{is.irl}
\alias{is.locs}
\alias{is.df}
\title{Convert between Locations, IRangesList, and Ranged Data Frame}
\usage{
locs2irl(locs, ...)

irl2locs(irl)

irl2df(irl)

df2irl(df)

locs2df(locs)

df2locs(df)

is.irl(x)

is.locs(x)

is.df(x)
}
\arguments{
\item{locs, irl, df}{object to be converted}

\item{...}{optional, add metadata to object}
}
\description{
Inspect, or convert back and forth between:
}
\details{
\itemize{
\item \code{locs}: list of matrices, output of \code{\link[stringr:str_locate]{stringr::str_locate_all()}}.
\item \code{irl}: \code{\link[IRanges:IRangesList-class]{IRangesList}} object,
\item \code{df}: ranged data frame, modified data frame version of \code{irl}
}
}
