% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{logistic.formula}
\alias{logistic.formula}
\title{...Title...}
\usage{
...usage.code...

...details...
}
\arguments{
\item{.param1.}{...param1.description...}

\item{.param2.}{...param2.description...}
}
\value{
...description.of.data.returned...
}
\description{
...Description...
}
\examples{
...examples.here....
}
\seealso{
\code{\link{cdiff.method}}
}
\author{
Ying Taur
}
\keyword{...}
\keyword{keyword1}
\keyword{keyword2}
