% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{logistic_trans_breaks}
\alias{logistic_trans_breaks}
\title{Breaks for Logistic Tranformation}
\usage{
logistic_trans_breaks(inner.range)
}
\arguments{
\item{inner.range}{the 2-value numeric used in logistic_trans}
}
\value{
break function returning break values.
}
\description{
Simple breaks for logistic, just use the inner.range used to define the curve.
}
