% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{logit}
\alias{logit}
\title{Logistic regression}
\usage{
logit(
  data,
  yvar,
  ...,
  return.model.obj = FALSE,
  firth = FALSE,
  formatted = TRUE
)
}
\arguments{
\item{data}{the data frame containing the variables to be analyzed.}

\item{yvar}{the outcome (bare unquoted).}

\item{...}{predictors in the model (bare unquoted).}

\item{return.model.obj}{if \code{TRUE}, returns the model object of the \code{coxph} command}

\item{formatted}{returns a formatted regression table (default \code{TRUE}). Otherwise, return the raw, unformatted regression table (essentially, the output of \code{broom::tidy}, plus a few additional columns)}

\item{starttime}{optional parameter specifying analysis start time.}
}
\value{
by default, returns a formatted regression table
}
\description{
Logistic regression
}
