% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{ls.object.sizes}
\alias{ls.object.sizes}
\title{Display sizes of objects in memory}
\usage{
ls.object.sizes(envir = .GlobalEnv)
}
\arguments{
\item{envir}{the environment (or object) from which to list objects. Default is \code{.GlobalEnv}}
}
\value{
a data frame showing objects and the object size, in Mb.
}
\description{
Use this to see what is occupying memory
}
