% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{make.competing.endpt}
\alias{make.competing.endpt}
\title{Make a competing endpoint}
\usage{
make.competing.endpt(data, newvar, primary, ..., censor = NULL)
}
\arguments{
\item{data}{the data to be modified, containing the endpoints to be combined}

\item{newvar}{the name (unquoted) of the new competing survival endpoint to be created (creates \code{newvar}, plus \code{paste0(newvar,"_day")}}

\item{primary}{the original survival endpoint, to be converted to a competing endpoint}

\item{...}{columns representing competing endpoints.}

\item{censor}{optional variable representing censoring times. Default is to use a censor of \code{Inf} time.}
}
\value{
Returns \code{data}, with a newly defined survival endpoint (\code{newvar}), which represents the combined competing endpoint.
\code{newvar} is the numeric indicator of the endpoint,
\code{newvar_day} is the survival time,
\code{newvar_code} is a character showing the value definition,
\code{newvar_info} shows the status of all endpoints, in order.
You would primarily use \code{newvar} and \code{newvar_day} with packages such as \code{cmprsk} for competing risk analysis.
}
\description{
Combine endpoints together, into a combined competing survival endpoint. This is to be used in competing risk analysis, such as with the cmprsk package.
}
\details{
Note, endpoints (primary and competing) can be specified either as a "varname" and "varname_day" pair, representing survival indicator and survival time,
or a single column representing positive endpoints (\code{NA} or \code{Inf}) otherwise.
}
\examples{
# create a combined endpoint
cid.patients \%>\% make.competing.endpt(competing.enterodom,enterodom30,dead,strepdom30,proteodom30,30)
}
\author{
Ying Taur
}
