% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{make.datetime}
\alias{make.datetime}
\title{Create Date-Time (POSIXct object)}
\usage{
make.datetime(date, time)
}
\arguments{
\item{date}{Date object}

\item{time}{character with time in it}
}
\value{
Returns POSIXct object with date and time combined.
}
\description{
Note: The particular use is in ytdata::visits() function.
}
