% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{make.names}
\alias{make.names}
\title{Make Syntactically Valid Names (Ying's version)}
\usage{
make.names(x, ...)
}
\arguments{
\item{names}{character vector to be coerced to syntactically valid names. This is coerced to character if necessary.}

\item{verbose}{logical indicating whether or not to display info name cleanup. Default is \code{FALSE}. Dataframe only}
}
\value{
Data frame with corrected names
}
\description{
Make syntactically valid names out of character vectors.
}
\details{
Like original \code{make.names}, but gets rid of any repeating periods('.'), as well as periods at the end. This is just an aesthetic modification.
}
