% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{make.surv.endpt}
\alias{make.surv.endpt}
\title{Make a survival endpoint}
\usage{
make.surv.endpt(data, newvar, primary, ..., censor = NULL, competing = FALSE)
}
\arguments{
\item{data}{the data to be modified, containing the endpoints to be combined}

\item{newvar}{the name (unquoted) of the new competing survival endpoint to be created (creates \code{newvar}, plus \code{paste0(newvar,"_day")})}

\item{primary}{the original survival endpoint, to be converted to a competing endpoint}

\item{...}{columns representing competing endpoints.}

\item{censor}{variable representing censoring times. Default is to use censoring times from the primary... or time=\code{Inf}, if it doesn't exist.}

\item{competing}{whether to code as competing. If FALSE, competing endpoints will be censored.}
}
\value{
Returns \code{data}, with a newly defined survival endpoint (\code{newvar}), which represents the combined competing endpoint.
\code{newvar} is the numeric indicator of the endpoint,
\code{newvar_day} is the survival time,
\code{newvar_code} is a character showing the value definition,
\code{newvar_info} shows the status of all endpoints, in order.
You would primarily use \code{newvar} and \code{newvar_day} with packages such as \code{cmprsk} for competing risk analysis.
}
\description{
Construct a survival endpoint, by specifying times or survival other survival endpoints.
These can be regular survival or competing endpoints (which you would analyze with something like Fine-Grey in \code{cmprsk} package).
}
\details{
Note, endpoints (primary and competing) can be specified either as a "varname" and "varname_day" pair, representing survival indicator and survival time,
or a single column representing positive endpoints (\code{NA} or \code{Inf}) otherwise.

If survival endpoints are specified, note that censored times may be ignored.
}
\examples{
# create a combined endpoint
cid.patients \%>\% make.endpt(competing.enterodom,enterodom30,dead,strepdom30,proteodom30,30)
}
\author{
Ying Taur
}
