% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{make.table}
\alias{make.table}
\title{Make Table}
\usage{
make.table(data, vars, by = NULL, showdenom = FALSE,
  fisher.test = TRUE)
}
\arguments{
\item{data}{Data frame containing data to be described.}

\item{vars}{character vector of variables within \code{data} to be summarized.}

\item{by}{Optional, variable name (character) by which to summarize the data. Each separate value will be a column of data in the table.}

\item{showdenom}{logical, whether to show denominator in the cells.}

\item{fisher.test}{fisher logical, whether or not to calculate Fisher exact tests. Only performed if \code{by} is also specified.}
}
\value{
Returns a data frame formatted to be summary table.
}
\description{
Creates a summary table (data frame) variables from the data.
}
\details{
This was written to create a "Table 1" of a manuscript.
}
\examples{
make.table(mtcars,c("cyl","gear"))
make.table(mtcars,c("cyl","gear"),by="vs",showdenom=TRUE)
}
\author{
Ying Taur
}
