% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{make.tax.legend}
\alias{make.tax.legend}
\title{Make tax legend}
\usage{
make.tax.legend(tax.palette = yt.palette2, fontsize = 5)
}
\arguments{
\item{tax.palette}{a list of formulas specifying the palette. Default is \code{yt.palette2}.}

\item{fontsize}{Font size. Default is 5.}
}
\value{
a ggplot object showing the legend.
}
\description{
Make tax legend
}
