% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{make_table}
\alias{make_table}
\title{Make Table}
\usage{
make_table(data, ..., by = NULL, denom = FALSE, maxgroups = 10, fisher = TRUE)
}
\arguments{
\item{data}{Data frame containing data to be described.}

\item{...}{column names (bare) within \code{data} to be summarized.}

\item{by}{optional variable name (bare) by which to summarize the data. Each separate value will be a column of data in the table.}

\item{denom}{whether to show the denominator in the summary}

\item{maxgroups}{max number of groups before collapsing into an "Other" category.}

\item{fisher}{whether or not to perform Fisher test. Performed if \code{by=...} is specified.}
}
\value{
Returns a data frame formatted to be summary table.
}
\description{
Creates a summary table (data frame) variables from the data.
}
\details{
This was written to create a "Table 1" of a manuscript.
}
\examples{
make_table(mtcars,cyl,gear)
}
\author{
Ying Taur
}
