% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{make_taxonomy_distinct}
\alias{make_taxonomy_distinct}
\title{Make taxonomy distinct}
\usage{
make_taxonomy_distinct(phy)
}
\arguments{
\item{phy}{phyloseq object}
}
\value{
modified phyloseq object with corrected names.
}
\description{
Modifies the tax_table of phyloseq object such that each level is distinct.
}
