% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{makeContent.taxfittexttree}
\alias{makeContent.taxfittexttree}
\title{makeContent.taxfittexttree}
\usage{
\method{makeContent}{taxfittexttree}(x)
}
\description{
called by: \code{GeomTaxonomy::draw_panel()} > \code{\link[=taxfittextGrob]{taxfittextGrob()}} for fitted tax labels
modified version of \code{ggfittext:::makeContent.fittexttree()}.
To save time rendering the labels, remove rows where height  is less than \code{unit(1,"char")}.
}
