% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{max2}
\alias{max2}
\alias{min2}
\title{Ying's Min/Max}
\usage{
max2(..., na.rm = FALSE)

min2(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric or character arguments.}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
A length-one vector representing max or min.
}
\description{
Similar to \code{min}/\code{max} command, except that if all values are \code{NA}s, the function returns
\code{NA} when \code{na.rm=TRUE} is specified, instead of \code{+/-Inf}.
}
\details{
This is useful when using the function repetitively and it's possible that everything can be \code{NA}.
This might come in handy if running min/max functions across an \code{apply} or \code{ddply} command.
}
\section{Functions}{
\itemize{
\item \code{max2}: \code{max2} maximum value.

\item \code{min2}: \code{max2} minimum value.
}}
\examples{
a <- c(1,2,3)
b <- c(1,2,NA)
c <- c(NA,NA,NA)

# produces same answer
max(a)
max2(a)
max(b)
max2(b)

max(b,na.rm=TRUE)
max2(b,na.rm=TRUE)

max(c)
max2(c)

#Produces different answer
max(c,na.rm=TRUE)
max2(c,na.rm=TRUE)
}
\author{
Ying Taur
}

