% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{middle.pattern}
\alias{middle.pattern}
\title{Middle Pattern}
\usage{
middle.pattern(start = "", middle = ".+", end = "")
}
\arguments{
\item{start}{Expression prior to middle expression.}

\item{middle}{Middle expression to be found.}

\item{end}{Expression after middle expression.}
}
\value{
A regular expression (Perl-style).
}
\description{
Creates a regular expression to extract text in between two patterns. This utilizes Perl-style lookahead and lookbehind assertions.
}
\examples{
text <- c("start[target]end","start[target 2]end","start[target 3]")
str_extract(text,middle.pattern("start",".+","end"))
str_extract(text,middle.pattern("start",".+"))
}
