% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{midpoint}
\alias{midpoint}
\title{Get Midpoint}
\usage{
midpoint(tstart, tstop)
}
\arguments{
\item{tstart}{the start value}

\item{tstop}{the stop value}
}
\value{
a vector of values represent the midpoint between \code{tstart} and \code{tstop}.
}
\description{
For 2 vectors, get the center. Works with numeric, Date, POSIXct.
}
\details{
This is useful when dealing with Date and POSIXct, since you can't just add them together and divide by 2.
}
