% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{mix.colors}
\alias{mix.colors}
\title{Mix colors together (hex format)}
\usage{
mix.colors(colors)
}
\arguments{
\item{colors}{a vector of colors}
}
\value{
A single color vector representing mixing of colors in the vector.
}
\description{
Mix colors together (hex format)
}
