% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{mutate.phyloseq}
\alias{mutate.phyloseq}
\title{Mutate phyloseq object}
\usage{
\method{mutate}{phyloseq}(phy, ...)
}
\arguments{
\item{phy}{phyloseq object}

\item{...}{mutate commands}
}
\description{
Mutate operations on phyloseq object
}
