% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{name_along}
\alias{name_along}
\title{Object Names}
\usage{
name_along(x)
}
\arguments{
\item{x}{R object}
}
\value{
A character vector
}
\description{
Similar to \code{\link[base:names]{base::names()}}, but always same length as the object, and will be \code{NA} if no name.
}
