% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{occurs.within}
\alias{occurs.within}
\title{Determines if tstart-tstop occurs anywhere within interval.}
\usage{
occurs.within(tstart, tstop, start.interval, stop.interval)
}
\description{
Determines if tstart-tstop occurs anywhere within interval.
}

