% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{ordering}
\alias{ordering}
\title{Describe order of vectors}
\usage{
ordering(...)
}
\arguments{
\item{...}{variables to inspect and order}
}
\value{
A character vector of same length as variables, showing ordering
}
\description{
Summarize the order of variables in a character vector.
}
\examples{
mtcars \%>\%
  mutate(ord=ordering(mpg,cyl,qsec,20)) \%>\%
  count(ord)
}
