% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{overlaps}
\alias{overlaps}
\alias{envelops}
\title{Determines if 2 sets of time intervals overlap.}
\usage{
overlaps(
  start1,
  stop1,
  start2,
  stop2,
  check = TRUE,
  start_NA = NA,
  stop_NA = NA
)

envelops(
  start1,
  stop1,
  start2,
  stop2,
  check = TRUE,
  start_NA = NA,
  stop_NA = NA
)
}
\arguments{
\item{start1}{start times for interval 1}

\item{stop1}{stop times for interval 1}

\item{start2}{start times for interval 2}

\item{stop2}{stop times for interval 2}

\item{check}{whether to check if start comes before stop.}

\item{start_NA}{start value if NA. Default is to leave as NA.}

\item{stop_NA}{stop value if NA. Default is to leave as NA.}
}
\description{
\code{overlaps()} determines if interval 1 and 2 overlap in any way.
\code{envelops()} determines if interval 1 contains interval 2 entirely.
}
\examples{
library(dplyr)
times <- tribble(
  ~subj, ~startA, ~stopA, ~startB, ~stopB,
  1,     1,       2,      3,       4,
  2,     1,       3,      2,       4,
  3,     1,       4,      2,       3,
  4,     2,       3,      1,       4,
  5,     2,       4,      1,       3,
  6,     3,       4,      1,       2
)
times <- times \%>\%
  mutate(overlaps=overlaps(startA,stopA,startB,stopB),
         envelops=envelops(startA,stopA,startB,stopB))

ggplot() +
  geom_segment(data=times,aes(x=startA,xend=stopA,y=subj,yend=subj,color="A"),size=1) +
  geom_segment(data=times,aes(x=startB,xend=stopB,y=subj+0.1,yend=subj+0.1,color="B"),size=1) +
  geom_text(data=times,aes(x=0.5,y=subj,label=paste("overlaps=\n",overlaps))) +
  geom_text(data=times,aes(x=0,y=subj,label=paste("envelops=\n",envelops)))
}
