% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{paste2}
\alias{paste2}
\title{Paste 2}
\usage{
paste2(..., sep = " ", collapse = NULL)
}
\arguments{
\item{...}{one or more \code{R} objects, to be converted to character vectors (same as \code{paste}).}

\item{sep}{a character string to separate the terms (same as \code{paste}).}

\item{an}{optional character string to separate the results (same as \code{paste}).}
}
\value{
A character vector of the concatenated values.
}
\description{
Similar to \code{paste} command, except that \code{NA}s are not converted to text.
If all fields are \code{NA}, then return \code{NA} if collapse if specified.
}
\details{
This is useful when dealing with \code{NA} values. \code{paste} produces character \code{"NA"} values,
of course I could just convert those to actual \code{NA} values afterwards. However, if I use \code{collapse} option,
I can frequently get a lot of character combinations with \code{NA} in it, where it can be hard to remove.
This function comes in handy when using the \code{collapse} option and I need to guarantee a single character value.
An example is the \code{fun.aggregate} function for \code{dcast}, which requires a function that returns a single value.
}
\examples{
a <- c(1,2,3)
b <- c(4,5,NA)
c <- c(NA,NA,NA)

# Produces same output
paste(a,collapse=",") #produces "1,2,3"
paste2(a,collapse=",") #produces "1,2,3"

# Output is different for these
paste(b,collapse=",") #produces "4,5,NA"
paste2(b,collapse=",") #produces "4,5"

paste(c,collapse=",") #produces "NA,NA,NA"
paste2(c,collapse=",") #produces NA

paste(a,b,collapse=",") #produces "1 4,2 5,3 NA"
paste2(a,b,collapse=",") #produces "1 4,2 5,3"
}
\author{
Ying Taur
}
