% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{phy.calc.relative.abundance}
\alias{phy.calc.relative.abundance}
\title{Transform phyloseq data to relative abundances}
\usage{
phy.calc.relative.abundance(phy)
}
\arguments{
\item{phy}{phyloseq object to be transformed}
}
\value{
phyloseq with relative abundances
}
\description{
Convenience function that simply runs \code{phyloseq::transform_sample_counts(phy, function(x) x / sum(x) )}
}
\examples{
otu <- cid.phy \%>\%
  phy.calc.relative.abundance() \%>\%
  filter(Phylum=="Proteobacteria",
         Patient_ID=="318",prune_unused_taxa = F) \%>\%
  get.otu.melt()
ggplot(otu,aes(x=sample,y=numseqs,fill=otu,label=Genus)) +
  geom_taxonomy(width = 0.7, show.ribbon = TRUE)
}
