% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{phy.collapse.bins.old}
\alias{phy.collapse.bins.old}
\title{Collapse phyloseq into smaller bins}
\usage{
\method{phy.collapse.bins}{old}(
  phy,
  level = length(rank_names(phy)),
  fillin.levels = FALSE,
  criteria = max.pctseqs <= 0.001 | pct.detectable <= 0.005
)
}
\arguments{
\item{phy}{phyloseq object to be collapsed}

\item{level}{number of tax levels to evaluate and collapse by, starting with \code{otu} and moving up. For instance, level=4 means collapse at otu, Species, Genus, Family.}

\item{fillin.levels}{Whether or not to fill in \code{NA} values with the collapsed taxa name. Default is \code{FALSE}.}

\item{criteria}{an expression that evaluates to TRUE/FALSE, whether to collapse at a particular level. Create this using calculated stats for each taxa (see Details)}
}
\value{
collapsed phyloseq object
}
\description{
Collapse phyloseq object into smaller bins, using specified criteria based on the data.
}
\details{
These variables are available for criteria, for each taxon:
\itemize{
\item \code{n.detectable} number of samples where taxa abundance is above 0.
\item \code{pct.detectable} percent of samples where taxa abundance is above 0.
\item \code{mean.pctseqs} mean relative abundance for a taxon.
\item \code{median.pctseqs} median relative abundance for a taxon.
\item \code{max.pctseqs} highest relative abundance of sequences for a taxon (across all samples)
\item \code{min.pctseqs} lowest relative abundance of sequences for a taxon (across all samples)
\item \code{total.numseqs} total number of sequences for a taxon (across all samples)
\item \code{max.numseqs} highest number of sequences for a taxon (across all samples)
\item \code{min.numseqs} lowest number of sequences for a taxon (across all samples)
\item \code{n.samps} total number of samples (regardless of abundance)
\item \code{n.rows}  total number of rows for a taxon
}
}
\examples{
phy.binned <- phy.collapse.bins(cid.phy)
phy.binned2 <- phy.collapse.bins(cid.phy, level=5,
                                 criteria=mean.pctseqs<0.001 & n.detectable<=2)
}
