% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{phy.collapse}
\alias{phy.collapse}
\title{Collapse Phyloseq Into Taxonomy}
\usage{
phy.collapse(
  phy,
  taxranks = c("Superkingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  short_taxa_names = TRUE
)
}
\arguments{
\item{phy}{A phylsoeq object.}

\item{taxranks}{tax levels to collapse by. Default is \code{c("Superkingdom","Phylum","Class","Order","Family","Genus","Species")}.}

\item{short_taxa_names}{How to name the collapsed OTUs. If \code{TRUE}, use name of first OTU plus number of OTUs being collapsed. If \code{FALSE}, paste the OTU names together.}
}
\value{
A phyloseq object with OTUs collapsed.
}
\description{
In a phyloseq object, combine OTUs of the same taxonomic classification.
}
\details{
Similar to \code{phyloseq::tax_glom}, but with the following differences:
(a) it performs much faster,
(b) requires all tax levels to be specified (instead of assuming all ranks to the left of the tax-level)
(c) the new OTU names will specify old OTU names separated by '|'
}
\examples{
library(phyloseq)
cid.phy.family <- phy.collapse(cid.phy,taxranks=c("Kingdom", "Phylum", "Class", "Order", "Family"))
cid.phy
cid.phy.family
}
