% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{phy.combine}
\alias{phy.combine}
\title{Combine phyloseq objects into one}
\usage{
phy.combine(x, y)
}
\arguments{
\item{x}{first phyloseq object}

\item{y}{second phyloseq object}
}
\value{
Merged phyloseq
}
\description{
Similar to \code{\link[phyloseq:merge_phyloseq]{phyloseq::merge_phyloseq()}}, but with different behavior and verbosity.
}
\details{
\code{phy.combine()} and \code{\link[phyloseq:merge_phyloseq]{phyloseq::merge_phyloseq()}} have many similarities:
\enumerate{
\item Both functions combine phyloseq objects, with the ability to have shared taxa and shared samples across phyloseqs,
using \code{taxa_names()} and \code{sample_names()} as scaffold.
\item If there is a mismatch in terms of columns in \code{tax_table()} or \code{sample_data()}, both functions
try to keep the data by inserting \code{NA} values.
}

However, there are a few important differences:
\enumerate{
\item When merging \code{otu_table()}, if both phyloseqs have overlapping sample-taxon fields,
\code{\link[phyloseq:merge_phyloseq]{phyloseq::merge_phyloseq()}} will add abundances together;
whereas \code{phy.combine()} will make sure every duplicated field is identifical in value, and then use that value.
\item When merging \code{sample_data()} or \code{tax_table()}, if the phyloseqs do not match in terms of a particular value,
\code{\link[phyloseq:merge_phyloseq]{phyloseq::merge_phyloseq()}} will just use the value from the first phyloseq, whereas
\code{phy.combine()} will throw an error.
\item If \code{sample_data()}, \code{tax_table()}, or \code{refseq()} only appears in one phyloseq and not the other,
\code{\link[phyloseq:merge_phyloseq]{phyloseq::merge_phyloseq()}} will remove any taxa or sample that is not represented fully.
\code{phy.combine()} will warn and then insert \code{NA} blanks for \code{sample_data()} and \code{tax_table()}, and
remove for \code{refseq()}.
\item If there is a \code{phy_tree()} present in either phyloseq object, \code{\link[phyloseq:merge_phyloseq]{phyloseq::merge_phyloseq()}} will fail completely;
\code{phy.combine()} will remove and issue warning.
\item \code{phy.combine()} will issues warnings if the two phyloseq objects do not have any shared taxa.
}
}
\examples{
# distinct samples with some taxa overlap (most straightforward)
# warns if there is phylo tree
phy1 <- cid.phy \%>\% filter(sample \%in\% sample[1:4])
phy2 <- cid.phy \%>\% filter(sample \%in\% sample[5:8])
phy.combine(phy1,phy2)

# overlapping samples (combines and checks overlap, issues warning)
phy1 <- cid.phy \%>\% filter(sample \%in\% sample[1:6])
phy2 <- cid.phy \%>\% filter(sample \%in\% sample[4:9])
phy.combine(phy1,phy2)

# incomplete columns in x or y (issues warning)
phy1 <- cid.phy \%>\% filter(sample \%in\% sample[1:4]) \%>\%
  select_tax_table(-Kingdom)
phy2 <- cid.phy \%>\% filter(sample \%in\% sample[5:8]) \%>\%
  select_sample_data(-Consistency)
phy.combine(phy1,phy2)

# completely missing sample_data, tax_table, refseq (issues warning)
phy1 <- cid.phy \%>\% filter(sample \%in\% sample[1:4]) \%>\%
  phyloseq(otu_table(.),phy_tree(.))
phy2 <- cid.phy \%>\% filter(sample \%in\% sample[5:8])
phy.combine(phy1,phy2)
}
