% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{phy.has.refseq}
\alias{phy.has.refseq}
\title{Check if phyloseq has refseq}
\usage{
phy.has.refseq(phy)
}
\arguments{
\item{phy}{phyloseq object to be checked}
}
\value{
Returns \code{TRUE} if phyloseq object has a refseq component
}
\description{
Check if phyloseq has refseq
}
\examples{
phy.has.refseq(cid.phy)
}
