% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{phy.ordinate}
\alias{phy.ordinate}
\title{Ordination on Phyloseq}
\usage{
phy.ordinate(phy, method, distance = NULL, ...)
}
\arguments{
\item{phy}{Phyloseq data to be calculated}

\item{method}{Ordination method, passed to \code{\link[phyloseq:ordinate]{phyloseq::ordinate()}}. But can also be \code{umap} or \code{tsne}.}

\item{distance}{Character string passed to \code{\link[=calc.distance]{calc.distance()}}, or distance object. Note that not all ordination methods need distance.}

\item{...}{Additional arguments passed to ordination methods.}
}
\value{
a list containing data and ordination object
}
\description{
Calculates ordination data. This is sort of a modified version of \code{\link[phyloseq:ordinate]{phyloseq::ordinate()}}, with some additional
methods such as \code{tsne} and \code{umap}, and other tweaks.
}
\details{
Ordination methods that use a distance matrix:
\itemize{
\item \code{NMDS}: (passed to \code{\link[vegan:metaMDS]{vegan::metaMDS()}} through \code{\link[phyloseq:ordinate]{phyloseq::ordinate()}})
\item \code{PCoA/MDS}: (passed to \code{\link[ape:pcoa]{ape::pcoa()}} through \code{\link[phyloseq:ordinate]{phyloseq::ordinate()}})
\item \code{CAP}: requires \code{formula} parameter, and only uses distance from \code{\link[vegan:vegdist]{vegan::vegdist()}} (passed to \code{\link[vegan:capscale]{vegan::capscale()}} through \code{\link[phyloseq:ordinate]{phyloseq::ordinate()}}).
\item \code{tsne}: try tuning the \code{perplexity} parameter, typically \code{5-50} (passed to \code{\link[Rtsne:Rtsne]{Rtsne::Rtsne()}})
}

Ordination methods with no distance matrix:
\itemize{
\item \code{RDA}:(passed to \code{\link[vegan:cca]{vegan::rda()}} through \code{\link[phyloseq:ordinate]{phyloseq::ordinate()}})
\item \code{CCA}: (passed to \code{\link[vegan:cca]{vegan::cca()}} through \code{\link[phyloseq:ordinate]{phyloseq::ordinate()}})
\item \code{DCA}: (passed to \code{\link[phyloseq:decorana]{phyloseq::decorana()}} through \code{\link[phyloseq:ordinate]{phyloseq::ordinate()}})
\item \code{DPCoA}: (passed to \code{\link[phyloseq:DPCoA]{phyloseq::DPCoA()}} through \code{\link[phyloseq:ordinate]{phyloseq::ordinate()}})
\item \code{umap}: (passed to \code{\link[umap:umap]{umap::umap()}})
}
}
\examples{
library(tidyverse)
library(patchwork)
library(rlang)
phy <- cid.phy \%>\%
  filter(otu \%in\% otu[1:20],prune_unused_samples = TRUE)
ord.list <- list(
  RDA=phy.ordinate(phy,method="RDA"),
  CCA=phy.ordinate(phy,method="CCA"),
  DCA=phy.ordinate(phy,method="DCA"),
  DPCoA=phy.ordinate(phy,method="DPCoA"),
  umap=phy.ordinate(phy,method="umap"),
  NMDS=phy.ordinate(phy,method="NMDS",distance="unfold.horn"),
  PCoA=phy.ordinate(phy,method="PCoA",distance="unfold.horn"),
  tsne=phy.ordinate(phy,method="tsne",distance="unfold.horn"),
  CAP=phy.ordinate(phy,method="CAP",distance="bray",formula=~Consistency)
)
glist <- ord.list \%>\% imap(~{
  ggplot(.x$data,aes(x=axis1,y=axis2,color=Consistency)) +
    geom_point() + theme(aspect.ratio=1) + ggtitle(.y)
})
inject(wrap_plots(!!!glist))
}
