% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{phy.prepare.ggtree}
\alias{phy.prepare.ggtree}
\title{Prepare data for ggtree plot}
\usage{
phy.prepare.ggtree(
  phy,
  layout = "circular",
  ...,
  ggtree = NULL,
  sortby = NULL,
  xmin.tip = 0.6,
  xring.range = c(1.05, 1.4)
)
}
\arguments{
\item{phy}{phyloseq as source of the data}

\item{layout}{layout of ggtree, passed to \code{\link[ggtree:ggtree]{ggtree::ggtree()}}}

\item{...}{other options, passed to \code{\link[ggtree:ggtree]{ggtree::ggtree()}}}

\item{ggtree}{ggtree object to be used. If not specified, one is created from phyloseq using options \code{layout} and \code{...} parameters.}

\item{sortby}{sort data by a column in \code{sample_data()}. This affects order of concentric circles, if plotting heatmap.}

\item{xmin.tip}{how far out will the lowest tip be, relative to the furthest tip? Default is \code{0.6}.}

\item{xring.range}{relative to the furthest tip, across what range will the sample data appear? Default is \code{c(1.05,1.4)}.}
}
\value{
A list of objects for ggtree plotting. Includes \code{otu}, \code{tax}, \code{samp}, \code{gd}, \code{ggtree}, \code{xdict}, \code{ydict},
which contain plotting information.
}
\description{
Prepares for ggtree plot by calculating things and making them available.
}
\examples{
library(dplyr)
library(phyloseq)
library(ggtree)
phy <- cid.phy \%>\%
  filter(Patient_ID==172) \%>\%
  mutate_sample_data(daylabel=str_glue("day {day}"))
gg <- phy.prepare.ggtree(phy,sortby=day)
gg$ggtree +
  geom_tile(data=gg$otu,aes(x=x,y=y,fill=otu,alpha=pctseqs),color="gray") +
  geom_point(data=gg$tax,aes(x=x,y=y)) +
  geom_segment(data=gg$tax,aes(x=x,xend=x.ring.min,y=y,yend=y),color="gray",linetype="dotted") +
  geom_text(data=gg$samp,aes(x=x,y=0,label=daylabel),angle=-90) +
  geom_text(data=gg$tax,aes(x=x.ring.max,y=y,label=Genus,hjust=hjust,angle=angle),color="dark gray") +
  scale_alpha_continuous(trans=log_epsilon_trans()) +
  scale_fill_taxonomy(data=gg$otu,fill=otu)
}
