% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{phy.unfold.taxranks}
\alias{phy.unfold.taxranks}
\title{Perform taxonomic unfolding on phyloseq data}
\usage{
phy.unfold.taxranks(phy, verbose = TRUE)
}
\arguments{
\item{phy}{phyloseq to be unfolded}

\item{verbose}{whether or not to print progress}
}
\value{
a modified phyloseq object
}
\description{
Taxonomic unfolding refers to converting all taxonomic levels into separate taxa rows in the OTU table.
Primarily used for making a distance metric more "taxonomy-aware".
Used by \code{\link[=calc.distance]{calc.distance()}} when \code{unfold=TRUE}.
}
\details{
Basically the operation runs \code{\link[=phy.collapse]{phy.collapse()}} at each tax level of \code{phy}, then binds them all together
in a single OTU table with taxa features at every level.
Note that sample counts will be elevated and will no longer reflect the sequencing depth.
}
