% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{phy.use.refseq.as.taxanames}
\alias{phy.use.refseq.as.taxanames}
\title{Use refseq as taxa names for a phyloseq}
\usage{
phy.use.refseq.as.taxanames(phy)
}
\arguments{
\item{phy}{phyloseq object to be modified}
}
\value{
modified phyloseq, with \code{refseq()} as \code{taxa_names()}.
}
\description{
Sets \code{taxa_names(phy)} to be \code{refseq(phy)}, if it exists.
}
\examples{
new.phy <- phy.use.refseq.as.taxanames(cid.phy)
head(taxa_names(cid.phy),3)
head(taxa_names(new.phy),3)
}
