% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{phy.use.tax.tree}
\alias{phy.use.tax.tree}
\title{Use phyloseq taxonomy table as phylo tree}
\usage{
phy.use.tax.tree(phy, collapse.singles = TRUE)
}
\arguments{
\item{phy}{phyloseq object to be modified (should contain \code{\link[phyloseq:tax_table-methods]{phyloseq::tax_table()}} element)}

\item{collapse.singles}{argument passed on to \code{\link[=as.phylo.formula]{as.phylo.formula()}}. Default is \code{TRUE}. Note that this is necessary for
making sure number of nodes in the tree does not exceed number of tips; turns out to be a problem if you are using
this to calculate Unifrac distances in \code{\link[=calc.distance]{calc.distance()}}}
}
\value{
a modified phyloseq object
}
\description{
Creates a tree from taxonomy (using \code{\link[=as.phylo.formula]{as.phylo.formula()}}), and uses that as the phyloseq object's
\code{\link[phyloseq:phy_tree-methods]{phyloseq::phy_tree()}} element.
}
\examples{
phy.use.tax.tree(cid.phy)
}
