% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{pivot_wider_recode}
\alias{pivot_wider_recode}
\title{Pivot data from long to wide, with recoding}
\usage{
pivot_wider_recode(
  data,
  id_cols = NULL,
  names_from = name,
  values_from = value,
  names_recodes,
  names_else.value = "other",
  names_sort = TRUE,
  values_sep = "::",
  values_fill = NULL,
  values_fn = ~paste(.x, collapse = "|"),
  unused_fn = NULL
)
}
\arguments{
\item{data}{data to be pivotted.}

\item{id_cols}{\code{\link[tidyr:tidyr_tidy_select]{tidy-select}} columns that identify each observation. Used in \code{\link[tidyr:pivot_wider]{tidyr::pivot_wider()}}.}

\item{names_from}{\code{\link[tidyr:tidyr_tidy_select]{tidy-select}} column names to be pivotted. Used in \code{\link[tidyr:pivot_wider]{tidyr::pivot_wider()}}.}

\item{values_from}{\code{\link[tidyr:tidyr_tidy_select]{tidy-select}} column values to be pivotted. Used in \code{\link[tidyr:pivot_wider]{tidyr::pivot_wider()}}.}

\item{names_recodes}{recodes to be done on \code{names_from} prior to pivotting. Used in \code{\link[=recode.grep]{recode.grep()}}.}

\item{names_else.value}{default value of recoding. Used in \code{\link[=recode.grep]{recode.grep()}}.}

\item{names_sort}{whether to sort columns by order of recode.grep. Default is \code{TRUE}}

\item{values_sep}{character separator between name and value. Default is \code{"::"}}

\item{values_fill}{value used when value is missing. Used in \code{\link[tidyr:pivot_wider]{tidyr::pivot_wider()}}.}

\item{values_fn}{How multiple values are combined. Default is \code{~paste(.x,collapse="|")}, which collapses into a single string. Used in \code{\link[tidyr:pivot_wider]{tidyr::pivot_wider()}}.}

\item{unused_fn}{A function performed on unused columns. Default is \code{NULL}.}
}
\value{
Pivotted data
}
\description{
Somewhat similar to \code{\link[tidyr:pivot_wider]{tidyr::pivot_wider()}}, where but where the names_from column is recoded using \code{recode.grep},
prior to pivotting. This is primarily useful for restructuring data that comes in long format (name-value pairs).
}
\details{
In addition to recoding, there are other changes:
\enumerate{
\item The names_from column is recoded using
\item asdf
}
}
\examples{
library(tidyverse)
cid.bsi \%>\%
  pivot_wider_recode(
    id_cols=Patient_ID,
    names_from=organism,
    values_from=organism,
    names_recodes=c("klebs|coli|serratia|pseudomonas|bacter|acineto|steno|lepto|neiss"="gram.neg",
                    "staphylococcus|CNST|VRE|enterococcus|bacillus|micrococc"="gram.pos",
                    "clostridium|fusobact|bacteroides"="anaerobe"),
    values_fill="<none>")
}
