% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{plot.pca}
\alias{plot.pca}
\title{Plot Principal Components Analysis}
\usage{
\method{plot}{pca}(dist, data = FALSE, prefix = NA)
}
\arguments{
\item{dist}{distance matrix to be plotted.}

\item{data}{logical, if \code{TRUE}, returns a data frame of PCA axes instead of the plot. Default is \code{FALSE}.}

\item{prefix}{character, an optional prefix text for PCA variable names. E.g. if \code{"unifrac"} is used, \code{"PCA1"} becomes \code{"unifrac.PCA1"}.}
}
\value{
Returns a \code{ggplot2} graph of PCA1 and PCA2.
}
\description{
Plots PCA from distance matrix data.
}
\examples{
xxx
}
\author{
Ying Taur
}
\keyword{...}
\keyword{keyword1}
\keyword{keyword2}

