% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{pluck2}
\alias{pluck2}
\alias{hoist2}
\title{Pluck/Hoist 2}
\usage{
pluck2(.x, ..., .simplify = TRUE)

hoist2(.data, .col, ..., .simplify = TRUE)
}
\arguments{
\item{.x}{list object to be accessed.}

\item{...}{list of accessors, similar to \code{\link[purrr:pluck]{purrr::pluck()}}.}

\item{.simplify}{Whether to simplify output. Default is \code{TRUE}.}
}
\value{
Retireved list items.
}
\description{
Similar to \code{\link[purrr:pluck]{purrr::pluck()}} and \code{\link[purrr:hoist]{purrr::hoist()}}, but with additional accessor methods.
}
\details{
Comparison with \code{\link[purrr:pluck]{purrr::pluck()}} and \code{\link[purrr:hoist]{purrr::hoist()}} accessors:
\itemize{
\item Like: Can use a character or numeric index
\item Unlike: Can use purrr-style formula or function that evaluates to logical.
\item Unlike: Can supply use logical or vector of integer indexes.
\item Unlike: Each accessor can retrieve multiple hits. Like anything with the same name.
}
}
