% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{pretty_scientific}
\alias{pretty_scientific}
\title{Fancy Scientific Notation}
\usage{
pretty_scientific(l, parse = TRUE)
}
\arguments{
\item{l}{number vector to be formatted.}
}
\value{
Expression for l, in scientific notation.
}
\description{
Use to format axes in ggplot with scientific notation.
}
\examples{
pretty_scientific(c(111e-12,230000022.11111,0.1234567))
ggplot(mtcars,aes(mpg*1e-6)) + geom_bar() + scale_x_continuous(label=pretty_scientific)
}
