% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{read.mothur.taxfile}
\alias{read.mothur.taxfile}
\title{Read in mothur taxonomy file.}
\usage{
read.mothur.taxfile(tax.file)
}
\arguments{
\item{tax.file}{mothur taxonomy file to be read.}
}
\value{
Returns a data frame containing taxonomy
}
\description{
Read in mothur taxonomy file.
}
\author{
Ying Taur
}

