% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{read.oligos}
\alias{read.oligos}
\title{Read in oligos file.}
\usage{
read.oligos(oligo.file)
}
\arguments{
\item{oligo.file}{oligo file to be read. If oligo.file is a directory, all oligo files (*.oligos) will be read in.}
}
\value{
Returns a data frame containing oligo information
}
\description{
Reads in oligos file, listing pertinent information
}
\examples{
...examples.here....
}
\author{
Ying Taur
}
\keyword{...}
\keyword{keyword1}
\keyword{keyword2}

