% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{read.otu.table}
\alias{read.otu.table}
\title{Read OTU table from text file}
\usage{
read.otu.table(otu.file, row.names = "OTUId")
}
\arguments{
\item{otu.file}{text file to be read, containing otu table data}

\item{otu.row.names}{specify column listing OTU names. This is passed to \code{read.delim};
i.e. can be vector of accual row names, single number of column, or character string name of the column.}
}
\value{
Dataframe containing otu table
}
\description{
Used for uparse pipeline to read in otu table.
}
\details{
Assumes tab-delimited file, with 'OTUID' to specify otu column.
}
