% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{read.tree.uparse}
\alias{read.tree.uparse}
\title{Read Tree File (Label-Fix)}
\usage{
read.tree.uparse(tree.file)
}
\arguments{
\item{tree.file}{Newick-formatted text file}
}
\value{
Returns a \code{phylo} object, with original labels.
}
\description{
Read in tree file created by uparse pipeline.
}
\details{
Functions like \code{read.tree} have trouble reading uparse-generated trees, because the OTU names contain semicolons(;), e.g. "\code{'OTU_2080;size=5;'}". This function temporarily replaces \code{;} with \code{__}, reads in successfully, and reverts back to names with semicolons. Also, removes the single quotes (\code{\'}) from the name. If read in this way, the tree can be quickly merged into a phyloseq object.
}
\examples{
b <- import_biom("uparse/total.8.otu-tax.biom")
tr <- read.tree.uparse("uparse/total.10.tree")
phy <- merge_phyloseq(b,tr)
}
\author{
Ying Taur
}

