% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{read_all_excel}
\alias{read_all_excel}
\title{Read Multiple Excel Sheets Into a List of Data Frames}
\usage{
read_all_excel(
  ...,
  col_names = TRUE,
  keep.nested = FALSE,
  bare.filename = TRUE
)
}
\arguments{
\item{...}{Either a file(s) or folder(s). If a folder is specified, it will look for all files ending in (.xlsx/.xls).}

\item{col_names}{\code{TRUE} to use the first row as column names, \code{FALSE} to get default names, or a character
vector giving a name for each column. This is passed to \code{\link[readxl:read_excel]{readxl::read_excel()}} function.}

\item{keep.nested}{If \code{TRUE}, returns a nested list of files and then sheets. Otherwise, a list of sheets is normally returned.}

\item{bare.filename}{Whether to use bare vs. full path as the filename.}
}
\value{
A named list of data frames, where each data frame represents a sheet.
}
\description{
Read Multiple Excel Sheets Into a List of Data Frames
}
