% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{read_excel2}
\alias{read_excel2}
\title{Read Excel File 2}
\usage{
read_excel2(
  path,
  sheet = 1,
  col_names = TRUE,
  col_types = NULL,
  na = "",
  skip = 0
)
}
\arguments{
\item{path}{Path to the xls/xlsx file}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an integer (the position of the sheet). Defaults to the first sheet.}

\item{col_names}{Either TRUE to use the first row as column names, FALSE to number columns sequentially from X1 to Xn, or a character vector giving a name for each column.}

\item{col_types}{Either NULL to guess from the spreadsheet or a character vector containing "blank", "numeric", "date" or "text". (YTmod: can be named vector listing only variables you want to change)}

\item{na}{Missing value. By default readxl converts blank cells to missing data. Set this value if you have used a sentinel value for missing values.}

\item{skip}{Number of rows to skip before reading any data.}
}
\value{
A data frame consisting of Excel data.
}
\description{
Same as \code{readxl::read_excel} function, but col_types can be named vector
}
