% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{regex.widget}
\alias{regex.widget}
\title{Regular Expression Widget}
\usage{
regex.widget(vec, port = 4567)
}
\arguments{
\item{vec}{a character vector where regex searches will be applied.}
}
\value{

}
\description{
A widget for exploring regular expressions on a vector
}
