% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{regex.widget0}
\alias{regex.widget0}
\title{Regular Expression Widget (old)}
\usage{
regex.widget0(vec, port = 4567)
}
\arguments{
\item{vec}{a character vector where regex searches will be applied.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
A widget for exploring regular expressions on a vector
}
