% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{remove.na.cols}
\alias{remove.na.cols}
\title{Remove NA columns}
\usage{
remove.na.cols(data, verbose = FALSE)
}
\arguments{
\item{data}{data frame to be filtered.}

\item{verbose}{logical indicating whether or not to display info on columns removed. Default is \code{FALSE}.}
}
\value{
The original data frame, with blank columns removed.
}
\description{
Remove NA columns
}
