% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{remove.na.rows}
\alias{remove.na.rows}
\title{Remove NA rows}
\usage{
remove.na.rows(data, verbose = FALSE)
}
\arguments{
\item{data}{data frame to be filtered.}

\item{verbose}{logical indicating whether or not to display info on rows removed. Default is \code{FALSE}.}
}
\value{
The original data frame, with blank rows removed.
}
\description{
Remove NA rows
}
