% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{round.data.frame}
\alias{round.data.frame}
\title{Rounding of Numbers}
\usage{
\method{round}{data.frame}(x, digits = 0)
}
\arguments{
\item{x}{data frame to be rounded}

\item{digits}{number of decimals to round}
}
\value{
the data frame, with numeric values rounded.
}
\description{
Performs rounding function on data frame.
}
\examples{
round(mtcars)
}
