% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{run.as.tempscript}
\alias{run.as.tempscript}
\title{Run As Temp Script}
\usage{
run.as.tempscript(expr)
}
\arguments{
\item{expr}{the code to be executed}

\item{env}{environment to be saved. Default is \code{parent.frame()}}
}
\description{
Use this to run code separately from the terminal, as an instance of R that is separate from the current one.
It saves an image of the environment (the variables needed to run the code),
loads any necessary packages, and
executes the code (runs Rscript of a temporary .R file), and loads results back into the current instance of R.
}
\details{
I use this as a quick fix measure if I come across code that doesn't execute well in my current R environment for some reason.
For example, for some reason I can't pull SQL data via RJDBC when running RStudio, for unclear reasons. So I execute the pull
code within this function.
}
