% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{run_r}
\alias{run_r}
\alias{run_r_bg}
\alias{run_r_callargs}
\title{Evaluate expression in another R session}
\usage{
run_r(expr, envir = parent.frame())

run_r_bg(expr, envir = parent.frame())

run_r_callargs(expr, envir = parent.frame())
}
\arguments{
\item{expr}{expression to be run in separate R session.}

\item{envir}{the environment to execute the code.}
}
\value{
For \code{run_r}: the value of the evaluated expression.
For \code{run_r_bg}: an \code{\link[callr:r_process]{callr::r_process}} object, which has a \code{\link[callr:get_result]{callr::get_result()}} method to collect the result.
For \code{run_r_callargs}: a list containing a modified function and a list of arguments to pass to the function;
this is designed to be the arguments that can be plugged into \code{\link[callr:r]{callr::r()}} or \code{\link[callr:r_bg]{callr::r_bg()}}.
}
\description{
Use these to run code in a separate instance of R, separate from your current console.
This is essentially doing what the \code{\link{callr}} package does, but just adds a few modifications to make it easier to use.
}
\details{
In some cases it is useful to run code in a separate R session... this is where functions from the \code{\link{callr}} package
come in handy, such as \code{\link[callr:r]{callr::r()}} or \code{\link[callr:r_bg]{callr::r_bg()}}. However, in order to use these, you have to be sure to:
\enumerate{
\item place the code inside an anonymous function,
\item refer to functions and variables explicitly from other packages using the :: notation,
\item pass any necessary local variables as arguments to the anonymous function.
}

\code{run_r()} and \code{run_r_bg()} are running \code{\link[callr:r]{callr::r()}} and \code{\link[callr:r_bg]{callr::r_bg()}}, except that you can
just insert the code without worrying about about the 3 modifications above.
}
