% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{sample_groups}
\alias{sample_groups}
\title{Sample n groups from a grouped table}
\usage{
sample_groups(grouped_df, size, weight = NULL, replace = FALSE)
}
\arguments{
\item{grouped_df}{a grouped data frame to be sampled}

\item{size}{number of groups to sample}

\item{weight}{sampling weights. Can be any expression that can be used in `summarize()` to calculate a valid weight.}

\item{replace}{sample with or without replacement? Default is `FALSE`}
}
\value{
a subset of the original grouped data frame
}
\description{
Sample n groups from a grouped table
}
\examples{
mtcars \%>\% group_by(gear) \%>\% sample_groups(2)
mtcars \%>\% group_by(gear) \%>\% sample_groups(5,weight=mean(mpg),replace=TRUE)
}
