% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{sample_n_groups_OLD}
\alias{sample_n_groups_OLD}
\title{Sample N Groups}
\usage{
sample_n_groups_OLD(grouped_df, size)
}
\arguments{
\item{grouped_df}{the grouped data frame to be sampled}

\item{size}{number of groups to sample}
}
\value{
a subset of the grouped data frame
}
\description{
Sample groups from a grouped data frame
}
\examples{
gdf <- mtcars \%>\% group_by(gear,carb)
sample_n_groups(gdf,3)
}
