% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{scale_fill_taxonomy}
\alias{scale_fill_taxonomy}
\alias{scale_color_taxonomy}
\title{Taxonomy color scales}
\usage{
scale_fill_taxonomy(
  ...,
  data,
  tax.palette = yt.palette3,
  fill = Species,
  aesthetics = "fill",
  guide = guide_taxonomy(keywidth = 3),
  drop = FALSE,
  na.value = "grey50"
)

scale_color_taxonomy(
  ...,
  data,
  tax.palette = yt.palette3,
  color = otu,
  aesthetics = "color",
  guide = guide_taxonomy(keywidth = 3),
  drop = FALSE,
  na.value = "grey50"
)
}
\arguments{
\item{data}{tax table to be used for determining colors (passed to \code{\link[=get.taxonomy.colordata]{get.taxonomy.colordata()}}). This
might be the same data used to generate the ggplot (but can be different).}

\item{tax.palette}{list of expressions as criteria for the palette (passed to \code{\link[=get.taxonomy.colordata]{get.taxonomy.colordata()}}).}

\item{fill}{the (bare unquoted) column of taxons by which colors will be assigned. This should typically be
the same as what you specify for \code{aes(fill=xxxxx)}, or at least contain the same values.
Default is \code{Species} (passed as \code{unitvar} to \code{\link[=get.taxonomy.colordata]{get.taxonomy.colordata()}}).}

\item{aesthetics}{The names of the aesthetics that this scale works with. Default is \code{"fill"}}

\item{guide}{function used to create a guide. Default is \code{guide_taxonomy(keywidth = 3)}.}

\item{drop}{Should unused factor levels be omitted from the scale? The default, \code{TRUE}, uses the levels that appear in the data; \code{FALSE} uses all the levels in the factor.}

\item{na.value}{The aesthetic value to use for missing (\code{NA}) values}
}
\description{
Discrete scales for taxonomy.
}
\details{
This is modified from \code{\link[ggplot:scale_fill_manual]{ggplot::scale_fill_manual()}}. It runs \code{\link[=get.taxonomy.colordata]{get.taxonomy.colordata()}} to obtain colors.
}
\examples{
otu <- prune_samples("179A",cid.phy) \%>\%
  phy.collapse.bins(level=2) \%>\%
  get.otu.melt() \%>\%
  arrange(Kingdom,Phylum,Class,Order,Family,Genus,taxon) \%>\%
  mutate(otu=fct_inorder(otu))

ggplot(otu,aes(x=otu,y=pctseqs,fill=otu)) +
  geom_col() +
  scale_y_continuous(trans=log_epsilon_trans(0.01)) +
  scale_fill_taxonomy(data=otu,fill=otu)

ggplot(otu,aes(x=otu,y=pctseqs,fill=otu)) +
  geom_col() +
  scale_y_continuous(trans=log_epsilon_trans(0.01)) +
  scale_fill_taxonomy(data=otu,fill=otu,
                      guide=guide_taxonomy(keywidth=1.5,
                                           keyheight=0.85,ncol=2))
}
