% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{search.data.columns}
\alias{search.data.columns}
\title{Search data columns for a Regex pattern}
\usage{
search.data.columns(data, pattern, ignore.case = TRUE)
}
\arguments{
\item{data}{data to be searched}

\item{pattern}{regex pattern to be searched}

\item{ignore.case}{whether to ignore case, default is \code{TRUE}}
}
\value{
a summary table of column hits
}
\description{
Use to search columns for a pattern.
}
