% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{select.phyloseq}
\alias{select.phyloseq}
\title{Select vars in phyloseq object}
\usage{
\method{select}{phyloseq}(phy, ...)
}
\arguments{
\item{phy}{phyloseq object}

\item{...}{select expressions}
}
\value{
phylseq object, modified
}
\description{
Perform \code{select} operation on phyloseq object. Can be operations to be done on \code{sample_data} or \code{tax_table}.
}
