% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{send.text.message}
\alias{send.text.message}
\title{Send a Text Message}
\usage{
send.text.message(message, number = "9149802489")
}
\arguments{
\item{message}{A string containing the message you want to send.}

\item{number}{The phone number you want to send the message to. Default is Ying's cell number.}
}
\value{
No value.
}
\description{
Send a text message to your phone!
}
\examples{
send.text.message("Hello, this is a text message")
}
\author{
Ying Taur
}

