% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{set.otu}
\alias{set.otu}
\title{Convert OTU table to phyloseq otu_table}
\usage{
set.otu(odata, taxa_are_rows = TRUE)
}
\arguments{
\item{odata}{otu table (matrix or dataframe with 'otu' column) to be converted back to otu_table.}
}
\value{
formatted \link[phyloseq:taxonomyTable-class]{tax_table}.
}
\description{
Use this on data.frames with tax data. The opposite of get.tax function. Make sure it contains the variable "otu".
}
