% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{set.samp}
\alias{set.samp}
\title{Convert data frame to phyloseq sample_data}
\usage{
set.samp(sdata)
}
\arguments{
\item{sdata}{Data frame to be converted back to sample_data}
}
\value{
formatted sample_data.
}
\description{
Use this on data with sample info. The opposite of function get.samp. Make sure it contains variable "sample"
}
