% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{set.tax}
\alias{set.tax}
\title{Convert data frame to phyloseq tax_table}
\usage{
set.tax(tdata)
}
\arguments{
\item{tdata}{dataframe to be converted back to tax_table.}
}
\value{
formatted tax_table.
}
\description{
Use this on data.frames with tax data. The opposite of get.tax function. Make sure it contains the variable "otu".
}

