% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{set_precedence}
\alias{set_precedence}
\title{Set Namespace Precedence}
\usage{
set_precedence(pkg)
}
\arguments{
\item{pkg}{package name to be prioritized}
}
\description{
Set Namespace Precedence
}
\examples{
library(dplyr)
library(plyr)

# will not work
mtcars \%>\% rename(mpg1=mpg)

set_precedence("dplyr")

# will work
mtcars \%>\% rename(mpg1=mpg)
}
