% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{shades}
\alias{shades}
\title{Color Shades}
\usage{
shades(color, ncolor = 3, variation = 1)
}
\arguments{
\item{color}{character, specifying the color you want to build shades around. (e.g. \code{"red"} or \code{"#1460fa"})}

\item{ncolor}{number specifying the length of the vector, i.e. how many different shades should be returned (default 3 shades).}

\item{variation}{a number from 0-1, which determines how different the shades will be. Smaller numbers will be more similar.}
}
\value{
Produces a character vector of colors, corresponding to shades of the specified color.
}
\description{
Creates different shades of the specified color.
}
\details{
Use this as a convenience function when creating your plots.
}
\examples{
sh1 <- shades("red",5)
show_col(sh1)

sh2 <- shades("red",20)
show_col(sh2)

sh3 <- shades("red",20,variation=0.5)
show_col(sh3)
}
\author{
Ying Taur
}

