% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{shell.exec}
\alias{shell.exec}
\title{Open a File using File Associations}
\usage{
shell.exec(file)
}
\arguments{
\item{file}{file to be opened}
}
\value{
No value.
}
\description{
Opens the specified file using the application specified.
}
\details{
Linux R doesn't have this function, but Windows R does. So this emulates on Linux using the 'xdg-open' terminal command..
}
\author{
Ying Taur
}

