% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{show.tax.legend}
\alias{show.tax.legend}
\title{Show Tax Legend}
\usage{
show.tax.legend(palette)
}
\arguments{
\item{palette}{palette (list of expressions) to be viewed}
}
\description{
Use this to preview the tax palette.
}
\examples{
show.tax.legend(bacterial.palette)
show.tax.legend(fungal.palette)
}
