% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\docType{data}
\name{StatTaxonomy}
\alias{StatTaxonomy}
\title{StatTaxonomy}
\format{
An object of class \code{StatTaxonomy} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatTaxonomy
}
\description{
The stat object for taxonomy data. This allows for transformed Y-axis.
}
\keyword{datasets}
