% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{stcox}
\alias{stcox}
\title{Cox Proportional Hazards Regression}
\usage{
stcox(..., starttime = "tstart", data, addto, as.survfit = FALSE,
  firth = TRUE, formatted = TRUE, logrank = FALSE, coxphf.obj = FALSE)
}
\arguments{
\item{.param1.}{...param1.description...}

\item{.param2.}{...param2.description...}
}
\value{
...description.of.data.returned...
}
\description{
Analyzes survival data by Cox regression.
}
\details{
Convenience function for survival analysis. Typically uses the \code{coxphf} function.
}
\examples{
...examples.here....
}
\author{
Ying Taur
}
\seealso{
\code{\link{cdiff.method}}
}
\keyword{...}
\keyword{keyword1}
\keyword{keyword2}

