% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{str_split_equal_parts}
\alias{str_split_equal_parts}
\title{Split String Into Approximately Equal Sized Parts}
\usage{
str_split_equal_parts(char, nparts = 2, sep = " ", collapse = "\\n")
}
\arguments{
\item{char}{character vector to be split.}

\item{nparts}{number of parts to split into. Default is \code{2}.}

\item{sep}{separator to split by. Default is a space, \code{" "}}

\item{collapse}{character to collapse by, after splitting. Default is \code{"\\n"}. Specify \code{NULL} to keep as a list.}
}
\value{
a character vector (or list) which has been split.
}
\description{
Split a string vector into separate lines of approximately equal size.
Typical use for this is formatting labels for plotting.
}
\examples{
str_split_equal_parts(char,4)
}
