% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{style_selection_4096}
\alias{style_selection_4096}
\title{Formats the highlighted selection such that no line exceeds 4096 characters.}
\usage{
style_selection_4096()
}
\value{

}
\description{
Formats the highlighted selection such that no line exceeds 4096 characters.
}
