% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{style_selection_oneline}
\alias{style_selection_oneline}
\title{Formats the highlighted selection into one line}
\usage{
style_selection_oneline()
}
\description{
Formats the highlighted selection into one line
}
