% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{summarize_tax_table}
\alias{summarize_tax_table}
\alias{summarize_sample_data}
\title{Add a Summary Column to \code{tax_table()} or \code{sample_data()}}
\usage{
summarize_tax_table(phy, ..., filter.zero = FALSE)

summarize_sample_data(phy, ..., filter.zero = FALSE)
}
\arguments{
\item{phy}{phyloseq to be modified}

\item{...}{name-value statements that would go into \code{dplyr::summarize()}}

\item{filter.zero}{whether or not to remove zero abundances (passed to \code{\link[=get.otu.melt]{get.otu.melt()}}).}
}
\value{
modified phyloseq
}
\description{
Modifies phyloseq object to contain a summary column to
one of the tables.
\code{summarize_tax_table()} adds a column to \code{tax_table()},
\code{summarize_sample_data()} adds a column to \code{sample_data()}
}
\details{
This is convenience function which is essentially running
\code{get.otu.melt() \%>\% group_by() \%>\% summarize(...)},
and then adding results to the table.
}
\examples{
phy.calcs <- cid.phy \%>\%
  summarize_tax_table(
    ever.dominated=any(pctseqs>0.3),
    n.samples.detectable=sum(numseqs>0)
  ) \%>\%
  summarize_sample_data(
    enterodom30=sum(pctseqs[Genus=="Enterococcus"])>0.3,
    predominant.taxa=taxon[which.max(numseqs)]
  )
}
