% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{tab}
\alias{tab}
\title{Tabulate}
\usage{
tab(var, sort = TRUE, pct = TRUE, as.char = FALSE, collapse = "\\n")
}
\arguments{
\item{var}{the vector to be tabulated}

\item{sort}{whether to sort results. Default \code{TRUE}}

\item{pct}{whether to display percents. Default \code{TRUE}}

\item{as.char}{logical specifying whether to return tabulation as a single character. Useful for summarizing data within grouping commands such as \code{ddply} or \code{group_by}/\code{summarize}}

\item{collapse}{if \code{as.char=TRUE}, how to collapse. Default \code{"\\n"}}
}
\value{
Returns a data frame with tabulations.
}
\description{
Tabulates frequencies of vectors. By default, sorts by frequency.
}
\examples{
tab(cid.patients$sex)
}
\author{
Ying Taur
}
