% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tax.plot}
\alias{tax.plot}
\title{Plot tax}
\usage{
tax.plot(
  t,
  xvar = sample,
  pctseqs = pctseqs,
  unitvar = Species,
  label = Species,
  tax.levels = c("Superkingdom", "Phylum", "Class", "Order", "Family", "Genus",
    "Species"),
  data = TRUE,
  label.pct.cutoff = 0.3
)
}
\arguments{
\item{t}{data frame containing melted tax data. Needs to have vars sample, pctseqs, Kingdom, ... , Species}

\item{xvar}{xvar by which to plot data. This should be a distinct identifier for samples.}

\item{pctseqs}{the relative abundance column name  (default \code{pctseqs})}

\item{unitvar}{unit variable (bare unquoted). Default \code{Species}}

\item{label}{column name (bare unquoted) for labels var. Default is \code{Species}.}

\item{tax.levels}{tax ranks.}

\item{data}{whether to return data frame only (\code{TRUE}), or  proceed with plotting (\code{FALSE}).}

\item{label.pct.cutoff}{cutoff by which to label abundances, stored in tax.label}
}
\value{
either ggplot2 object, or data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\author{
Ying Taur
}
