% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{taxfittextGrob}
\alias{taxfittextGrob}
\title{taxfittextGrob}
\usage{
taxfittextGrob(
  data,
  panel_scales,
  coord,
  padding.x = grid::unit(1, "mm"),
  padding.y = grid::unit(1, "mm"),
  min.size = 4,
  grow = FALSE,
  reflow = FALSE,
  hjust = NULL,
  vjust = NULL,
  fullheight = NULL,
  width = NULL,
  height = NULL,
  formatter = NULL,
  contrast = FALSE,
  place = "centre",
  outside = FALSE
)
}
\description{
called by: \code{GeomTaxonomy::draw_panel()} for fitted tax labels
modified version of \code{ggfittext:::GeomFitText$draw_panel()}
}
