% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{taxnonomy_scale}
\alias{taxnonomy_scale}
\title{Taxonomy scale constructor}
\usage{
taxnonomy_scale(
  aesthetic,
  data,
  tax.palette,
  unitvar,
  guide = guide_taxonomy(keywidth = 3),
  drop = FALSE,
  na.value = "grey50",
  na.translate = TRUE,
  expand = waiver(),
  name = waiver(),
  position = "left",
  limits = NULL
)
}
\description{
Similar to  \code{\link[ggplot2:::manual_scale]{ggplot2::::manual_scale()}}
}
