% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{timeit}
\alias{timeit}
\title{Time It Benchmarking}
\usage{
timeit(expr)
}
\arguments{
\item{expr}{expression to be benchmarked}
}
\description{
Times each step of code.
}
\examples{
timeit(1+2)
timeit({
  mt <- mtcars \%>\%
    select(-disp) \%>\%
    mutate(mpg1=mpg+1,
           cyl1=cyl+2) \%>\%
    slice(1:9)
})
}
