% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{univariate.logit}
\alias{univariate.logit}
\title{Univariate and Multivariate Cox Regression}
\usage{
univariate.logit(
  data,
  yvar,
  ...,
  multi = TRUE,
  multi.cutoff = 0.25,
  firth = FALSE,
  formatted = TRUE
)
}
\arguments{
\item{data}{the data frame containing the variables to be analyzed.}

\item{yvar}{the time-to-event outcome (bare unquoted).}

\item{...}{predictors in the model (bare unquoted). If a predictor time-dependent, the split the corresonding rows of the data frame.}

\item{multi}{if \code{TRUE}, perform multivariate analysis.}

\item{multi.cutoff}{P-value threshold for inclusion into the multivariate model (default is \code{0.25})}

\item{formatted}{returns a formatted regression table (default \code{TRUE}). Otherwise, return the raw, unformatted regression table (essentially, the output of \code{\link[broom:reexports]{broom::tidy()}}, plus a few additional columns)}

\item{starttime}{optional parameter specifying analysis start time.}
}
\value{
by default, returns a formatted regression table
}
\description{
Uses the \code{cox} function to perform a univariate and multivariate model analysis.
}
