% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{univariate.stcox}
\alias{univariate.stcox}
\title{Univariate Cox Proportional Hazards}
\usage{
univariate.stcox(
  yvar,
  xvars,
  starttime = "tstart",
  data,
  firth = TRUE,
  multi = FALSE,
  multi.cutoff = 0.2
)
}
\arguments{
\item{xvars}{column names of predictors.}

\item{data}{the data frame to be analyzed}

\item{firth}{whether to perform Firth's penalized likelihood correction.}

\item{multi}{whether to perform multivariate modelling of signficant univariate predictors}

\item{multi.cutoff}{if multivariate is done, the P-value cutoff for inclusion into the multivariate model.}

\item{yvars}{column name of survival endpoint.}

\item{tstart}{column name of start variable.}
}
\value{
A regression table containing results
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Perform univariate survival, then multivariate on significant variables.
}
