% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{view.dist}
\alias{view.dist}
\title{View Distance Metric}
\usage{
view.dist(dist, show.numbers = FALSE)
}
\arguments{
\item{dist}{distance metric to be viewed}

\item{show.numbers}{whether or not to show values.}
}
\description{
Usethis to quicklyview a distance metric.
}
\examples{
view.dist(dist(mtcars))
}
