% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{view.distance.matrix}
\alias{view.distance.matrix}
\title{View Distance Metric}
\usage{
view.distance.matrix(dist, show.numbers = FALSE)
}
\arguments{
\item{dist}{distance metric to be viewed}

\item{show.numbers}{whether or not to show values.}
}
\value{
a ggplot of the distance matrix
}
\description{
Usethis to quicklyview a distance metric.
}
\examples{
view.distance.matrix(dist(mtcars))
}
