% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{write_all_excel}
\alias{write_all_excel}
\title{Write multiple data frames to an Excel file}
\usage{
write_all_excel(..., file)
}
\arguments{
\item{...}{objects to be written to the Excel file. Can be a data frames or lists of data frames.}

\item{file}{the Excel file to be written}
}
\description{
Write multiple data frames to an Excel file
}
