% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\docType{package}
\name{yingtools2}
\alias{yingtools2}
\alias{yingtools2-package}
\title{Useful R Functions by Ying}
\description{
Ying's assorted collection of R functions. Primarily useful for manipulation, analysis, and plotting of data, particularly microbiome data.
}
\details{
To install/update this package, run: \code{devtools::install_github("ying14/yingtools2")}
}
\author{
Ying Taur

Maintainer: Ying Taur
}

