% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\docType{data}
\name{yt.palette3}
\alias{yt.palette3}
\title{YT Palette 3}
\format{
An object of class \code{list} of length 11.
}
\usage{
yt.palette3
}
\description{
The customary palette for Bacteria.
}
\keyword{datasets}
