% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{as.Date2}
\alias{as.Date2}
\title{Convert to Date (Ying's version)}
\usage{
as.Date2(vec)
}
\arguments{
\item{vec}{The vector to be converted.}
}
\value{
Returns vector converted to Date. If date+time, convert to POSIXct.
}
\description{
Examines strings and convert to Date format if necessary.
}
\examples{
#Access dates read by RODBC:
access.dates <- c("2009-012-03 00:00:00","2010-08-21 00:00:00","2013-07-01 00:00:00","2014-10-31 00:00:00")
as.Date2(access.dates)
}

