% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{calc.pairwise}
\alias{calc.pairwise}
\title{Calculate distances for specified samples from phyloseq}
\usage{
calc.pairwise(sample1, sample2, phy, method = "bray")
}
\arguments{
\item{sample1}{a character vector specifying the first sample(s) for comparison. Should be a sample in \code{phy}.}

\item{sample2}{a character vector specifying the second sample(s) for comparison. Should be a sample in \code{phy}.}

\item{phy}{a phyloseq object containing the samples to be compared.}

\item{method}{the distance metric method to be used. Can be a method from
\code{\link[phyloseq:distanceMethodList]{phyloseq}}, or \verb{"taxhorn'}.}
}
\value{
a vector of distances, corresponding to \code{sample1} and \code{sample2}.
}
\description{
Calculate distances from a phyloseq object. This is the similar to \code{\link[=calc.distance]{calc.distance()}},
except that this does not return a distance matrix. Instead, it returns a vector of distances,
only calculating the comparisons you specified.
}
\examples{
library(tidyverse)
tbl <- tibble(sample1=c("191A", "228A", "132A", "1045", "179B"),
              sample2=c("198A", "205B", "202C", "175B", "192D"))
tbl.dists <- tbl \%>\%
  mutate(taxhorn.dist=calc.pairwise(sample1,sample2,cid.phy,method="taxhorn"))
}
