% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{chop.endpoint}
\alias{chop.endpoint}
\title{Chop survival endpoint}
\usage{
chop.endpoint(data, newvar, oldvar, ..., censor.as.tdvar = FALSE)
}
\arguments{
\item{data}{the data frame with survival data}

\item{newvar}{the name (unquoted) of the new survival endpoint to be created (creates \code{newvar}, plus \code{paste0(newvar,"_day")}}

\item{oldvar}{the original survival endpoint, to be censored.}

\item{...}{columns representing censoring times.}

\item{censor.as.tdvar}{whether to censor endpoints occurring exactly at the censoring time. Use \code{TRUE} for time-dependent predictors, \code{FALSE} for endpoints.}
}
\value{
Returns \code{data}, with a newly defined survival endpoint (\code{newvar}), which has been censored wherever the censoring times occur before the original end of survival time.
}
\description{
For a given survival endpoint, censor at earlier timepoints, if they occur.
}
\examples{
# create a endpoint(dead30d), which represents death within 30 days or discharge.
new.pt <- cid.patients \%>\% chop.endpoint(dead30d,dead,30,discharge.day)
new.pt \%>\% select(dead30d, dead30d_day,dead, dead_day, discharge.day) \%>\% head(10)
}
\author{
Ying Taur
}
