% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\docType{data}
\name{cid.colors}
\alias{cid.colors}
\title{The color scheme used in CID manuscript (\link{https://academic.oup.com/cid/article/55/7/905/428203}).}
\format{
An object of class \code{character} of length 21.
}
\usage{
cid.colors
}
\description{
The color scheme used in CID manuscript (\link{https://academic.oup.com/cid/article/55/7/905/428203}).
}
\author{
Ying Taur
}
\keyword{datasets}
