% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cid94}
\alias{cid94}
\title{CID Data (94 patients)}
\format{Several data frames.}
\usage{
data(cid94)
}
\description{
A dataset containing clinical and microbiota data on the 94 patients from CID 2012.
}
\details{
\itemize{
  \item cid.patients Patient data (N=94)
  \item cid.meds Medications dispensed
  \item cid.bsi Bloodstream infection data
  \item cid.cdiff C.difficile testing
  \item cid.hosp Hospitalizations
  \item cid.phy 16S data (phyloseq format)
}
}
\keyword{datasets}
