% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{compare}
\alias{compare}
\alias{compare.default}
\alias{compare.character}
\alias{compare.data.frame}
\title{Compare objects}
\usage{
compare(x, ...)

\method{compare}{default}(x, y)

\method{compare}{character}(x, y)

\method{compare}{data.frame}(x, y, by = NULL)
}
\arguments{
\item{x}{object (vector or data frame) to be compared}

\item{y}{object (vector or data frame) to be compared}

\item{by}{variable(s) to join by and compare (for data frames only)}
}
\value{
displays a report of the comparison, and invisibly returns a table of comparison details.
}
\description{
Compare two objects (vector or data frame), and displays a summary of similarities and differences.
}
\examples{
library(tidyverse)
vec <- stringr::sentences[1:100]
compare(vec,vec)
compare(vec,rev(vec))
compare(vec,rep(vec,2))
compare(rep(vec,2),vec)
compare(vec,rep(vec[1:10],2))
compare(vec,sample(vec,size=200,replace=TRUE))
compare(vec[1:50],vec[51:100])

m <- mtcars \%>\% rownames_to_column("car")
m2 <- bind_rows(m,m)
m3 <- bind_rows(m,m,m)
m.sort <- m \%>\% arrange(mpg)
m.single <- m; m.single[4,4] <- 200
m.col <- m; m.col$disp <- 101
m.top <- m \%>\% slice(1:25)
m.bottom <- m \%>\% slice(10:n())
m.extracol <- m \%>\% mutate(mpg2=mpg+1,hp2=hp*100)
compare(m,m)
compare(m,m,by="car")
compare(m,m2,by="car")
compare(m2,m3,by="car")
compare(m,m,by="cyl")
compare(m,m.sort,by="car")
compare(m,m.single,by="car")
compare(m,m.single)
compare(m,m.col,by="car")
compare(m,m.top,by="car")
compare(m.top,m.bottom,by="car")
}
