% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{convert.dates}
\alias{convert.dates}
\title{Search and Convert Date variables}
\usage{
convert.dates(data, verbose = FALSE)
}
\arguments{
\item{data}{The data frame to be converted.}

\item{verbose}{logical indicating whether or not to display info on date conversions. Default is \code{FALSE}.}
}
\description{
In a given data frame, look for variables resembling dates and convert them to Dates.
}
\details{
Basically applies \code{as.Date2} to all variables.
}

