% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{copy.as.Rcode}
\alias{copy.as.Rcode}
\title{Convert object to R-code.}
\usage{
copy.as.Rcode(
  x,
  copy.clipboard = TRUE,
  fit = TRUE,
  width = getOption("width") - 15
)
}
\arguments{
\item{x}{object to be converted to R-code. Can be vector or data frame.}

\item{copy.clipboard}{logical, if \code{TRUE}, will copy the R-code to the Clipboard.}
}
\value{
Returns the R-code.
}
\description{
Produces R-code that would create the object inputted. I use this if I have some data object that I obtained
somehow but just want to declare it in the code.
}
\examples{
x <- c("a","b","c")
copy.as.Rcode(x)
x <- tibble("a"=1:4,"_b"=c(T,F,T,F),"c c"=Sys.Date()+4:1,"d"=factor(LETTERS[1:4]))
copy.as.Rcode(x)
}
\author{
Ying Taur
}
