% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{copy.to.clipboard}
\alias{copy.to.clipboard}
\alias{copy.to.clipboard.gg}
\title{Copy to Clipboard}
\usage{
copy.to.clipboard(x)

copy.to.clipboard.gg(
  obj,
  width = 10,
  height = 7,
  dpi = 150,
  pointsize = 12,
  rescale = "R"
)
}
\arguments{
\item{obj}{object to by copied. Can be data frame, matrix, table, vector.}
}
\description{
Copies object to the clipboard, which can be used to paste into other programs such as Word or Excel.
}
\details{
This is now done using the \code{clipr} package. Previously I did this manually for each operating system.
}
\author{
Ying Taur
}
