% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{cox}
\alias{cox}
\title{Cox Proportional Hazards Regression (TAKE 2)}
\usage{
cox(data, yvar, ..., starttime = NULL, return.split.data = FALSE,
  args5 = list(cens.model = "cox", model = "fg"))
}
\description{
STILL WRITING THIS
}
