% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{createSurvivalFrame}
\alias{createSurvivalFrame}
\title{Create Survival Data Frame}
\usage{
createSurvivalFrame(f.survfit, time0 = 0)
}
\arguments{
\item{f.survfit}{the survival data}

\item{time0}{time0 can be specified. It must be <= first event}
}
\value{
Returns survival data frame
}
\description{
Create Survival Data Frame
}
\examples{
...examples.here....
}
\author{
Ying Taur
}
\keyword{...}
\keyword{keyword1}
\keyword{keyword2}

