% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{filter.phyloseq}
\alias{filter.phyloseq}
\title{Filter phyloseq object}
\usage{
\method{filter}{phyloseq}(phy, ..., prune_unused_taxa = TRUE)
}
\arguments{
\item{phy}{phylseq object}

\item{...}{filter criteria}

\item{prune_unused_taxa}{whether or not to remove unused taxa. Default is \code{TRUE}}
}
\description{
Subset a phyloseq using \code{dplyr}-style.
}
