% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{geom_bracket}
\alias{geom_bracket}
\title{Draw Brackets}
\usage{
geom_bracket(
  mapping = NULL,
  data = NULL,
  stat = StatBracket,
  position = "identity",
  ...,
  na.rm = FALSE,
  bracket.width = unit(0.75, "char"),
  tip = "round",
  arrow = grid::arrow(angle = 20, length = bracket.width, ends = "both", type = "closed"),
  arrow.fill = NULL,
  lineend = "butt",
  linejoin = "round",
  parse = FALSE,
  check_overlap = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{bracket.width}{size of bracket edges. Default is \code{unit(0.75,'char')}.}

\item{tip}{The style of bracket to use. Can be one of: \verb{"round", "bare", "square", "ibeam", "arrow", "iarrow", "curly"}. Default is \code{"round"}.}

\item{arrow}{The style of arrow to use, if \code{tip="arrow"} or \code{tip="iarrow"}. Default selection is based on \code{bracket.width}.}

\item{arrow.fill}{Fill color to use for closed arrowheads, if \code{tip="arrow"} or \code{tip="iarrow"}.}

\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{parse}{If \code{TRUE}, the labels will be parsed into expressions and
displayed as described in \code{?plotmath}.}

\item{check_overlap}{If \code{TRUE}, text that overlaps previous text in the
same layer will not be plotted. \code{check_overlap} happens at draw time and in
the order of the data. Therefore data should be arranged by the label
column before calling \code{geom_text()}. Note that this argument is not
supported by \code{geom_label()}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{}
}
\description{
Draw brackets in a ggplot.
}
\details{
This draws brackets by using grid-drawking methods from \code{\link[=geom_text]{geom_text()}}, \code{\link[=geom_segment]{geom_segment()}}, and \code{\link[=geom_curve]{geom_curve()}}.
}
\section{Aesthetics}{

\code{geom_bracket()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{\link[=aes_position]{x}}}
\item \strong{\code{\link[=aes_position]{y}}}
\item \strong{\code{\link[=aes_position]{xend}}}
\item \strong{\code{\link[=aes_position]{yend}}}
\item \code{\link[=aes_colour_fill_alpha]{alpha}}
\item \code{angle}
\item \code{\link[=aes_colour_fill_alpha]{colour}}
\item \code{family}
\item \code{flip}
\item \code{fontface}
\item \code{fontsize}
\item \code{\link[=aes_group_order]{group}}
\item \code{hjust}
\item \code{label}
\item \code{lineheight}
\item \code{\link[=aes_linetype_size_shape]{linetype}}
\item \code{\link[=aes_linetype_size_shape]{linewidth}}
\item \code{vjust}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}

\examples{
library(tidyverse)
mt <- mtcars \%>\%
  mutate(mpg.group=cut2(mpg, n.splits=3, lvls=c("Low MPG", "Med MPG", "High MPG")))
g <- ggplot(mt) +
  geom_point(aes(x=mpg, y=hp, color=mpg.group), size=3)

# whole-data brackets
g + geom_bracket(aes(x=mpg, y=350), label="MPG values") +
  geom_bracket(aes(x=35, y=hp), label="HP values", flip=TRUE)

# brackets by group
g + geom_bracket(aes(x=mpg, y=350, label=mpg.group))

# adjust the brackets with padding
g + geom_bracket(aes(x=mpg - 0.25, xend=mpg + 0.25, y=350, label=mpg.group))

# brackets by group, vary height
g + geom_bracket(aes(x=mpg, y=ave(hp,mpg.group,FUN=max)+20, yend=ave(hp,mpg.group,FUN=max)+20, label=mpg.group))

# manually specified brackets
g + geom_bracket(aes(x=18, y=225, yend=350, label="Tier 1"),flip=TRUE) +
  geom_bracket(aes(x=18, y=220, xend=25,yend=150, label="Tier 2")) +
  geom_bracket(aes(x=26,xend=35, y=150, label="Tier 3"))

# customization
g + geom_bracket(aes(x=27, y=200, yend=350, label="200+ HP"), flip=TRUE,
  linewidth=2, color="purple", fontsize=7, angle=0, hjust=0, fontface="italic", bracket.width=unit(0.5,"in"))

# various tip styles
ggplot() +
  geom_bracket(aes(x=1, xend=5, y=7, label="round"), tip="round") +
  geom_bracket(aes(x=1, xend=5, y=6, label="bare"), tip="bare") +
  geom_bracket(aes(x=1, xend=5, y=5, label="square"), tip="square") +
  geom_bracket(aes(x=1, xend=5, y=4, label="ibeam"), tip="ibeam") +
  geom_bracket(aes(x=1, xend=5, y=3, label="arrow"), tip="arrow") +
  geom_bracket(aes(x=1, xend=5, y=2, label="iarrow"), tip="iarrow") +
  geom_bracket(aes(x=1, xend=5, y=1, label="curly"), tip="curly") +
  xlim(0,6)
}
