% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{geom_logrank}
\alias{geom_logrank}
\title{Generate label for Log-rank test results in ggplot2}
\usage{
geom_logrank(yvar, xvar, data, starttime = "tstart", x.pos, y.pos,
  logrank.fontsize = 5)
}
\arguments{
\item{.param1.}{...param1.description...}

\item{.param2.}{...param2.description...}
}
\value{
...description.of.data.returned...
}
\description{
Adds the p-value for a log-rank test to a ggplot2 graph.
}
\details{
...details...
}
\examples{
...examples.here....
}
\author{
Ying Taur
}
\seealso{
\code{\link{cdiff.method}}
}
\keyword{...}
\keyword{keyword1}
\keyword{keyword2}

