% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{geom_logrank}
\alias{geom_logrank}
\title{Generate label for Log-rank test results in ggplot2}
\usage{
geom_logrank(yvar, xvar, data, starttime = "tstart", pos, logrank.fontsize = 5)
}
\description{
Adds the p-value for a log-rank test to a ggplot2 graph.
}
\author{
Ying Taur
}
