% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{geom_taxonomy}
\alias{geom_taxonomy}
\title{Stacked Taxonomy Bar}
\usage{
geom_taxonomy(
  mapping = NULL,
  data = NULL,
  stat = StatTaxonomy,
  position = "stack",
  ...,
  label.pct.cutoff = 0.3,
  label.split = FALSE,
  width = 0.95,
  na.rm = FALSE,
  parse = FALSE,
  tax.palette = yt.palette3,
  drop = FALSE,
  fit.text = TRUE,
  reflow = FALSE,
  contrast = FALSE,
  check_overlap = FALSE,
  show.ribbon = FALSE,
  ribbon.alpha = 0.25,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{label.pct.cutoff}{cutoff abundance by which to label abundances. Default is 0.3. Only applies if \code{fit.text=FALSE}.}

\item{label.split}{whether to split label into two lines, using \code{\link[=str_split_equal_parts]{str_split_equal_parts()}}}

\item{width}{Bar width. By default, set to 0.95.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{parse}{The same argument used in \code{\link[=geom_text]{geom_text()}}. Only applies if \code{fit.text=FALSE}.}

\item{tax.palette}{a list of formulas used to assign colors. Each element should take the form: \verb{"<label>" = <true/false expression> ~ <color vector>}. See \code{\link[=scale_fill_taxonomy]{scale_fill_taxonomy()}} for examples and details.}

\item{fit.text}{whether or not the tax labels will be auto-fitted. \code{TRUE}: use \code{\link[ggfittext:geom_fit_text]{ggfittext::geom_fit_text()}} (the default); \code{FALSE}: use \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}.}

\item{reflow}{Whether or not to reflow the text. Default is \code{FALSE}. Only applies if \code{fit.text=TRUE}.}

\item{contrast}{Whether or not to vary font color based on background fill. Default is \code{FALSE}. Only applies if \code{fit.text=TRUE}.}

\item{check_overlap}{The same argument used in \code{\link[=geom_text]{geom_text()}}. This is only applicable if if \code{fit.text=FALSE}.}

\item{show.ribbon}{whether or not to display transition lines in the background. Default is \code{FALSE}.}

\item{ribbon.alpha}{alpha value of the ribbon. Default is \code{0.25}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
Creates stacked taxonomy barplots.
}
\details{
This geom is similar to \code{\link[ggplot:geom_col]{ggplot::geom_col()}}, where you can draw stacked barplots.
However there are a few additional capabilities:
\enumerate{
\item You can specify the aesthetic \code{label} to write text labels inside the bars.
The text labels can be further tweaked using aesthetics (\code{fontsize}, \code{fontcolour}, \code{fontface}, \code{fontalpha}),
or other parameters (\code{fit.text}, \code{reflow}, \code{contrast}, \code{label.split}, \code{parse}, \code{check_overlap})
\item If Y-axis is transformed, the total Y bar height will be correct.
In \code{\link[ggplot:geom_col]{ggplot::geom_col()}}, stacking of individual transformed Y values often leads to strange results.
Here, the transformed Y totals for each bar is calculated and plotted.
Fill colors occupy the barspace proportionally.
}
}
\section{Aesthetics}{

\code{geom_taxonomy()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \code{alpha}
\item \code{angle}
\item \code{colour}
\item \code{family}
\item \code{fill}
\item \code{fontalpha}
\item \code{fontcolour}
\item \code{fontface}
\item \code{fontsize}
\item \code{group}
\item \code{hjust}
\item \code{label}
\item \code{lineheight}
\item \code{linetype}
\item \code{linewidth}
\item \code{sample.colour}
\item \code{sample.linetype}
\item \code{sample.linewidth}
\item \code{vjust}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}

\examples{
library(tidyverse)
otu <- cid.phy \%>\%
  get.otu.melt() \%>\%
  filter(Patient_ID=="179")

# by sample
ggplot(data=otu,aes(x=sample,y=pctseqs,fill=otu,label=Genus)) +
  geom_taxonomy()

# by day
ggplot(data=otu,aes(x=day,y=pctseqs,fill=otu,label=Genus)) +
  geom_taxonomy(width=2)

# absolute abundance
ggplot(data=otu,aes(x=day,y=numseqs,fill=otu,label=Genus)) +
  geom_taxonomy(width=2)

# absolute abundance
ggplot(data=otu,aes(x=day,y=numseqs,fill=otu,label=Genus)) +
  geom_taxonomy(width=2) +
  scale_y_continuous(trans=log_epsilon_trans(100))

# if Y is transformed, colors are distributed proportionally
# (i.e. does not get messed up)
ggplot(data=otu,aes(x=day,y=numseqs,fill=otu,label=Genus)) +
  geom_taxonomy(width=2) +
  scale_y_continuous(trans=log_epsilon_trans(100))

# show ribbon transitions
ggplot(data=otu,aes(x=day,y=pctseqs,fill=otu,label=Genus)) +
  geom_taxonomy(width=2, show.ribbon = TRUE)
}
