% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{get.abundance}
\alias{get.abundance}
\title{Calculate Abundance from Phyloseq}
\usage{
get.abundance(phy, ..., counts = FALSE)
}
\arguments{
\item{phy}{\code{\link[phyloseq:phyloseq-class]{phyloseq}} object to be analyzed}

\item{...}{one or more taxonomic expressions defining the abundance to be calculated. Can be named, in order to specify the column name.}

\item{counts}{if \code{TRUE}, will return count rather than relative abundance}
}
\value{
a data frame containing sample identifier and abundance columns
}
\description{
Given a \code{\link[phyloseq:phyloseq-class]{phyloseq}}, calculate relative abundance for each sample.
}
\examples{
library(phyloseq)
get.abundance(cid.phy,pct.entero=Genus \%in\% "Enterococcus",pct.proteo=Phylum \%in\% "Proteobacteria")
}
