% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{get.otu}
\alias{get.otu}
\title{Extract Phyloseq otu_table}
\usage{
get.otu(phy, as.df = FALSE)
}
\arguments{
\item{phy}{phyloseq object containing otu_data}

\item{as.df}{if \code{TRUE}, return data frame (instead of matrix)}
}
\value{
Dataframe containing otu table
}
\description{
Creates data.frame from otu_table, storing the rownames as variable "otu". The opposite of set.tax function.
}
