% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{get.samp}
\alias{get.samp}
\title{Extract Phyloseq sample_data}
\usage{
get.samp(phy, stats = FALSE, measures = c("Observed", "InvSimpson", "Shannon"))
}
\arguments{
\item{phy}{phyloseq object containing \link[phyloseq:sample_data-class]{sample_data}}

\item{stats}{logical, whether or not to include summary statistics of samples. Stores \code{nseqs}, and diversity metrics.}

\item{measures}{diversity measures to calculate, if stats is \code{TRUE.} Default: \code{c("Observed","InvSimpson","Shannon")}}
}
\value{
Data frame containing \link[phyloseq:sample_data-class]{sample_data} data.
}
\description{
Returns \link[phyloseq:sample_data-class]{sample_data} component from phyloseq object, as a data frame.
}
\details{
This basically is similar to the function \code{\link[phyloseq:sample_data-methods]{phyloseq::sample_data()}}, but does a few extra things.
\enumerate{
\item Converts to a data frame
\item The sample name is stored in a column called \code{sample}. (\code{\link[phyloseq:phyloseq-class]{phyloseq}} normally stores as a row name)
\item Calculates number of sequences and alpha diversity metrics, if desired.
}

This function is the opposite of \code{\link[=set.samp]{set.samp()}}, which converts the data frame back into a \link[phyloseq:sample_data-class]{sample_data}.
Note that if the \code{\link[phyloseq:phyloseq-class]{phyloseq}} object does not contain \link[phyloseq:sample_data-class]{sample_data}, a data frame containing a single column, \code{sample}, is returned.
}
\examples{
get.samp(cid.phy)
}
