% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get.yt.palette}
\alias{get.yt.palette}
\title{Get YT Palette}
\usage{
get.yt.palette(tax, use.cid.colors = TRUE)
}
\arguments{
\item{tax}{either a data.frame, phyloseq, or tax_table}

\item{use.cid.colors}{whether to use classic CID colors}
}
\value{
a color palette that can be used in \code{ggplot2}
}
\description{
Get YT Palette
}
\author{
Ying Taur
}
