% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{gg.stack}
\alias{gg.stack}
\title{Stack and line up ggplot objects in a column}
\usage{
gg.stack(..., heights = NULL, gg.extras = NULL, gap = 0, margin = 1,
  units = "inches", as.list = FALSE)
}
\arguments{
\item{heights}{a numeric vector representing the relative height of each plot. Passed directly to \code{grid.arrange}.}

\item{gg.extras}{a list of ggplot objects that will be applied to all plots. Default is \code{NULL}.}

\item{gap}{size of gap between stacked plots. Default is 0}

\item{margin}{size of the margin around the plots. Default is 1.}

\item{units}{specifies units used for gap and margin.}
}
\value{
break function returning break values.
}
\description{
Use this to arrange ggplot objects, where the axes, plot, and legend are lined up correctly.
}
\details{
Performs these steps:
(1) change margins so that plots are closer together
(2) alters widths of each component so that the plots will line up nicely
(3) calls \code{grid.arrange(...,ncol=1)}
}
