% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{group_by_time_streaks}
\alias{group_by_time_streaks}
\title{Group by Time Streaks}
\usage{
group_by_time_streaks(data, time, indicator, ..., gap = Inf,
  na.skip = FALSE, add = FALSE)
}
\arguments{
\item{data}{data frame}

\item{time}{time variable}

\item{indicator}{variable to group consecutive streaks}

\item{...}{other variables to group by. These will be applied prior to grouping by time streaks.}

\item{gap}{time periods differing by this gap or less will be combined in the grouping variable. Default is \code{Inf}, i.e. no gap.}

\item{na.skip}{whether to ignore \code{NA} values in the indicator. Default is \code{FALSE}, which will just break streaks and provide a warning if they are encountered.}

\item{add}{Same as the add option in \code{group_by}. When TRUE, will add to groups, rather than overriding them.}
}
\value{
Returns \code{data}, but grouped by time streaks
}
\description{
Group time data by consecutive streaks of a certain indicator variable.
}
\details{
Similar to \code{group_by_time}, but for a different purpose. This function groups by consecutive values of the indicator variable.
This is to measure how long the indicator remains in the same state.
}
\author{
Ying Taur
}
