% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{guide_geom.taxonomy}
\alias{guide_geom.taxonomy}
\title{guide_geom.taxonomy}
\usage{
\method{guide_geom}{taxonomy}(guide, layers, default_mapping)
}
\description{
Called by name, by \code{\link[=guide_taxonomy]{guide_taxonomy()}}.
}
