% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{guide_taxonomy}
\alias{guide_taxonomy}
\title{Taxonomy Guide}
\usage{
guide_taxonomy(
  title = waiver(),
  override.tax.palette = NULL,
  title.position = NULL,
  title.theme = NULL,
  title.hjust = NULL,
  title.vjust = NULL,
  label = TRUE,
  label.position = NULL,
  label.theme = NULL,
  label.hjust = NULL,
  label.vjust = NULL,
  keywidth = 3,
  keyheight = NULL,
  direction = NULL,
  default.unit = "line",
  override.aes = list(),
  nrow = NULL,
  ncol = NULL,
  byrow = FALSE,
  reverse = FALSE,
  order = 0,
  ...
)
}
\description{
Used by \code{\link[=scale_fill_taxonomy]{scale_fill_taxonomy()}} to create a custom taxonomy legend.
Adapted from \code{\link[ggplot2:guide_legend]{ggplot2::guide_legend()}}.
}
