% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\docType{data}
\name{GuideTaxonomy}
\alias{GuideTaxonomy}
\title{GuideTaxonomy}
\format{
An object of class \code{GuideTaxonomy} (inherits from \code{GuideLegend}, \code{Guide}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GuideTaxonomy
}
\description{
ggproto method for tax legend.
}
\keyword{datasets}
