% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{hilight.clade}
\alias{hilight.clade}
\title{Highlight a clade in ggtree}
\usage{
hilight.clade(
  gdata,
  var = NULL,
  value = NULL,
  criteria = NULL,
  ymin = -Inf,
  ymax = Inf,
  label = "{value}\\n({var})",
  fill.color = NA,
  line.color = "dark gray",
  alpha = 1,
  xmax = NULL,
  xscalar = 1.5,
  fill.in = FALSE,
  font.size = 4
)
}
\arguments{
\item{gdata}{a data frame from a ggtree object.}

\item{var}{the variable to be matched (unquoted).}

\item{value}{the value that \code{var} needed for inclusion in the clade to be highlighted.}

\item{criteria}{an expression that evaluates to logical, can use as an alternative (or in addition to) var/value,}

\item{ymin}{the min value of y in the clade.}

\item{ymax}{the max value of y in the clade.}

\item{label}{A \code{glue} expression for the clade label. Default is \verb{"\{value}\n({var})"}}

\item{fill.color}{the clade fill color (default is \code{NA}, no fill)}

\item{line.color}{the color of the outline around the clade (default \code{"dark gray"})}

\item{alpha}{the alpha value of the fill color (default 1)}

\item{xmax}{the x depth at which the edge of the clade is drawn (optional)}

\item{xscalar}{if \code{xmax} is not specified, the x depth can be specified as a scaled factor of maximum x of all clade tips. Default is \code{1.5}}

\item{fill.in}{whether to fill all gaps in the clade (default is \code{FALSE})}

\item{font.size}{font size of the text (default is 4)}
}
\description{
Highlight a clade in ggtree
}
