% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{inner_join_replace}
\alias{inner_join_replace}
\alias{left_join_replace}
\alias{right_join_replace}
\alias{full_join_replace}
\title{Inner/Left/Right/Full Join with Replace}
\usage{
inner_join_replace(x, y, by = NULL, errorIfDifferent = TRUE)

left_join_replace(x, y, by = NULL, errorIfDifferent = FALSE)

right_join_replace(x, y, by = NULL, errorIfDifferent = FALSE)

full_join_replace(x, y, by = NULL, errorIfDifferent = FALSE)
}
\arguments{
\item{errorIfDifferent}{whether to throw an error if a difference is detected (default is \code{TRUE})}
}
\description{
Same as \code{inner_join}, \code{left_join}, \code{right_join}, and \code{full_join} in the \code{dplyr} package, except that variables with the
same column name will not be renamed with the ".x" and ".y" suffix.
Instead, the variables will be turned into one column if the variables are equal. If they are not equal, an error (or warning) is thrown.
}
\details{
This is a convenience function that just avoids the renaming of columns.
}
