% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{is.between}
\alias{is.between}
\title{Determines if x is between start and stop/}
\usage{
is.between(x, start, stop, check = TRUE)
}
\arguments{
\item{x}{vector of values to be checked}

\item{start}{vector of start time(s)}

\item{stop}{vector of stop time(s)}

\item{check}{whether to check if start comes before stop.}
}
\description{
Similar to \code{\link[dplyr:between]{dplyr::between()}}, except that the vectors are recycled, so x can be a fixed value.
}
