% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{is.same}
\alias{is.same}
\alias{all.same}
\alias{\%==\%}
\alias{\%!=\%}
\title{Test if 2 vectors are equal, element-wise}
\usage{
is.same(x, ...)

all.same(x, ...)

x \%==\% y

x \%!=\% y
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}
}
\value{
logical vector
}
\description{
Similar to \code{x == y}, but handles \code{NA} equality
}
\examples{
x <- c("A","B","C",NA,NA)
y <- c("A","X",NA,"D",NA)
# this is c(TRUE, FALSE, NA, NA, NA)
x == y
# but we want c(TRUE, FALSE, FALSE, FALSE, TRUE)
is.same(x,y)
x \%==\% y
all.same(x,y)
}
