% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{is.same}
\alias{is.same}
\alias{is.same.default}
\alias{is.same.data.frame}
\alias{\%==\%}
\alias{\%!=\%}
\title{Test if 2 objects are equal, element-wise}
\usage{
is.same(x, ...)

\method{is.same}{default}(x, y)

\method{is.same}{data.frame}(x, y)

x \%==\% y

x \%!=\% y
}
\arguments{
\item{x}{first object}

\item{y}{second object}
}
\value{
logical vector
}
\description{
Sort of like \code{x == y}, but handles \code{NA} equality,
and compares by name, if they exist and match.
}
\details{
Length of objects should be the same.
}
\examples{
x <- c("A","B","C",NA,NA)
y <- c("A","X",NA,"D",NA)
# this is c(TRUE, FALSE, NA, NA, NA)
x == y
# but we want c(TRUE, FALSE, FALSE, FALSE, TRUE)
is.same(x,y)
x \%==\% y
}
