% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{left_join_replace}
\alias{left_join_replace}
\title{Left Join Replace}
\usage{
left_join_replace(x, y, by = NULL, verbose = FALSE)
}
\arguments{
\item{x, }{y tbls to join}

\item{by}{a character vector of variables to join by.}

\item{verbose}{whether to comment on rows altered (default \code{FALSE})}
}
\value{
Returns a joined data frame
}
\description{
Like left_join, but replaces column names in x with new columns from y.
}
\author{
Ying Taur
}
