% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{log_epsilon_trans_breaks}
\alias{log_epsilon_trans_breaks}
\title{Breaks for Log Epsilon Tranformation}
\usage{
log_epsilon_trans_breaks(epsilon)
}
\arguments{
\item{epsilon}{scaling parameter used in \code{log_epsilon_trans}}
}
\value{
break function returning break values.
}
\description{
This is used by scant_trans as default method for breaks. Will fill in logs of 10.
}
