% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{make_taxonomy_distinct}
\alias{make_taxonomy_distinct}
\alias{make_taxonomy_distinct.phyloseq}
\alias{make_taxonomy_distinct.data.frame}
\title{Make taxonomy distinct}
\usage{
make_taxonomy_distinct(x, ...)

\method{make_taxonomy_distinct}{phyloseq}(phy, add.rank = FALSE)

\method{make_taxonomy_distinct}{data.frame}(
  data,
  taxranks = c("Superkingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  add.rank = FALSE
)
}
\arguments{
\item{phy}{\code{\link[phyloseq:phyloseq-class]{phyloseq}} object}

\item{add.rank}{logical, whether to add rank to taxon names: e.g. \code{Enterococcus} would be renamed to \code{Enterococcus (Genus)}. Default is \code{FALSE}}

\item{data}{data to be modified}

\item{taxranks}{vector of column names to be checked and modified.}
}
\value{
modified data frame or \code{\link[phyloseq:phyloseq-class]{phyloseq}} object with corrected names.
}
\description{
Modifies (if necessary) the data frame or \link[phyloseq:taxonomyTable-class]{tax_table} of \code{\link[phyloseq:phyloseq-class]{phyloseq}} object such that each rank level is a distinct identifier.
}
\details{
Sometimes there are two taxonomy naming issues that can cause issues or confusion:
\enumerate{
\item Two or more distinct taxonomies can have the same duplicate names at the lowest level, e.g. Genus=Ileibacterium can
either have Family=\code{"Erysipelotrichaceae" }or Family=\code{"Clostridiales Family XIII. Incertae Sedis"}.
\item The same name is used for 2 different levels, e.g. \code{"Actinobacteria"} is both a Phylum and a Class
}

This will handle by adding \verb{"#1", "#2", ...} to the name in the event of #1, and adds rank for #2.
}
