% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{mutate.phyloseq}
\alias{mutate.phyloseq}
\alias{mutate_sample_data}
\alias{mutate_tax_table}
\alias{filter.phyloseq}
\alias{filter_sample_data}
\alias{filter_tax_table}
\alias{select.phyloseq}
\alias{select_tax_table}
\alias{select_sample_data}
\title{Mutate/Filter/Select for phyloseq object}
\usage{
\method{mutate}{phyloseq}(phy, ..., verbose = FALSE)

mutate_sample_data(phy, ..., verbose = FALSE)

mutate_tax_table(phy, ..., verbose = FALSE)

\method{filter}{phyloseq}(
  phy,
  ...,
  prune_unused_taxa = TRUE,
  prune_unused_samples = FALSE,
  verbose = FALSE
)

filter_sample_data(phy, ..., prune_unused_taxa = TRUE, verbose = FALSE)

filter_tax_table(phy, ..., prune_unused_samples = FALSE, verbose = FALSE)

\method{select}{phyloseq}(phy, ..., verbose = FALSE)

select_tax_table(phy, ..., verbose = FALSE)

select_sample_data(phy, ..., verbose = FALSE)
}
\arguments{
\item{phy}{phylseq object to be modified}

\item{...}{expressions to be performed in \code{mutate}/\code{filter}/\code{select}.}

\item{verbose}{whether or not to display info (default \code{FALSE})}

\item{prune_unused_taxa}{whether or not to remove unused taxa after sample filtering. Default is \code{TRUE}}

\item{prune_unused_samples}{whether or not to remove unused samples after taxa filtering. Default is \code{FALSE}}
}
\description{
Manipulate \code{phyloseq} object using \code{dplyr}-style commands.
Operations can be performed on either \code{sample_data} or \code{tax_table},
and will automatically determine which one to modify (and will warn/error if it can't tell)
}
\details{
In these functions, each expression's variables are evaluated, to see if they should be done in \code{sample_data} or \code{tax_table}.
If an expression contains variables belonging to both, it will complain and raise an error.
If an expression does not contain variables that can be ascribed, it will raise a warning, and run the expression on \code{sample_data}.
Note that \code{mutate} and \code{filter} evaluates each expression in \code{...} separately, and \code{select} will evaluate \code{...} as a whole.
}
\examples{
cid.phy \%>\%
  mutate(sample_id=paste("Sample",Patient_ID),
         genus=paste(Genus,"(Genus)")) \%>\%
  filter(nseqs<1500,
         Consistency=="liquid") \%>\%
  select(-taxon,-Genus)
}
