% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{overlaps}
\alias{overlaps}
\title{Determines if 2 sets of time intervals overlap.}
\usage{
overlaps(start1, stop1, start2, stop2, check = TRUE)
}
\arguments{
\item{start1}{start times for interval 1}

\item{stop1}{stop times for interval 1}

\item{start2}{start times for interval 2}

\item{stop2}{stop times for interval 2}

\item{check}{whether to check if start comes before stop.}
}
\description{
Determines if 2 sets of time intervals overlap.
}
