% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{pca.plot}
\alias{pca.plot}
\title{Plot Principal Components Analysis}
\usage{
pca.plot(dist, data = FALSE, prefix = NA)
}
\arguments{
\item{dist}{distance matrix to be plotted.}

\item{data}{logical, if \code{TRUE}, returns a data frame of PCA axes instead of the plot. Default is \code{FALSE}.}

\item{prefix}{character, an optional prefix text for PCA variable names. E.g. if \code{"unifrac"} is used, \code{"PCA1"} becomes \code{"unifrac.PCA1"}.}
}
\value{
Returns a \code{ggplot2} graph of PCA1 and PCA2.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Plots PCA from distance matrix data.
}
\author{
Ying Taur
}
