% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{phy.collapse}
\alias{phy.collapse}
\title{Collapse Phyloseq Into Taxonomy}
\usage{
phy.collapse(phy, taxranks = rank_names(phy), short_taxa_names = TRUE)
}
\arguments{
\item{phy}{A phylsoeq object.}

\item{taxranks}{tax levels to collapse by. Default is \code{c("Superkingdom","Phylum","Class","Order","Family","Genus","Species")}.}

\item{short_taxa_names}{How to name the collapsed OTUs. If \code{TRUE}, use name of first OTU plus number of OTUs being collapsed. If \code{FALSE}, paste the OTU names together.}
}
\value{
A \code{\link[phyloseq:phyloseq-class]{phyloseq}} object with OTUs collapsed.
}
\description{
In a \code{\link[phyloseq:phyloseq-class]{phyloseq}} object, combine OTUs of the same taxonomic classification.
}
\details{
Similar to \code{\link[phyloseq:tax_glom]{phyloseq::tax_glom()}}, but with the following differences:
\enumerate{
\item it performs much faster,
\item requires all tax levels to be specified (instead of assuming all ranks to the left of the tax-level)
\item the new OTU names will specify old OTU names separated by \code{'|'}
}
}
\examples{
library(phyloseq)
cid.phy.family <- phy.collapse(cid.phy,taxranks=c("Kingdom", "Phylum", "Class", "Order", "Family"))
cid.phy
cid.phy.family
}
