% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{pretty_number}
\alias{pretty_number}
\title{Pretty Numeric Format (Non-scientific)}
\usage{
pretty_number(l, digits = 2)
}
\arguments{
\item{l}{number vector to be formatted.}
}
\value{
Expression for l, in non-scientific notation.
}
\description{
Use to format axes in ggplot with non-scientific notation. Good for abundances!
}
\examples{
pretty_number(c(111e-12,230000022.11111,0.1234567))
ggplot(mtcars,aes(mpg*1e-6)) + geom_bar() + scale_x_continuous(label=pretty_number)
}

