% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{prune_unused_taxa}
\alias{prune_unused_taxa}
\title{Prune Unused Taxa from a phyloseq object}
\usage{
prune_unused_taxa(phy)
}
\arguments{
\item{phy}{phyloseq object}
}
\value{
a phyloseq object, with empty taxa removed.
}
\description{
In a phyloseq object, remove any taxa that are not used in the samples.
}
