% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{read.clipboard}
\alias{read.clipboard}
\title{Read Clipboard}
\usage{
read.clipboard(sep = "\\n")
}
\arguments{
\item{sep}{separator between lines}
}
\value{
Contents of clipboard
}
\description{
Read clipboard into vector or data frame.
}
\author{
Ying Taur
}

