% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{read.oligos}
\alias{read.oligos}
\title{Read in oligos file.}
\usage{
read.oligos(oligo.file, remove.commented = TRUE)
}
\arguments{
\item{oligo.file}{oligo file to be read. If oligo.file is a directory, all oligo files (*.oligos) will be read in.}

\item{remove.commented}{logical indicating whether or not to remove commented lines (default TRUE)}
}
\value{
Returns a data frame containing oligo information
}
\description{
Reads in oligos file, listing pertinent information
}
\author{
Ying Taur
}
