% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{read.uparse.data}
\alias{read.uparse.data}
\title{Import UPARSE pipeline data and create phyloseq}
\usage{
read.uparse.data(dirpath)
}
\arguments{
\item{dirpath}{directory path (character) specifying folder where uparse data is.}
}
\value{
phyloseq object containing the specified UPARSE data.
}
\description{
Reads folder and looks for key files used to create phyloseq object.
}
\author{
Ying Taur
}
