% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{right.angle}
\alias{right.angle}
\title{Right Angle for Text on circular ggtree}
\usage{
right.angle(angle, hjust = FALSE)
}
\arguments{
\item{angle}{}

\item{hjust}{logical indicating whether to output the corresponding hjsut parameter.}
}
\value{
angle, turned 90 degrees.
}
\description{
Right Angle for Text on circular ggtree
}

