% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{select2}
\alias{select2}
\title{Select 2}
\usage{
select2(data, ...)
}
\arguments{
\item{data}{Data frame}

\item{...}{Comma separated list of unquoted expressions. You can treat variable names like they are positions. Use positive values to select variables; use negative values to drop variables.}
}
\value{
Returns \code{data}, but grouped by times and other variables.
}
\description{
Basically \code{dplyr::select}, but ignores variables that aren't found in the data frame.
}
\author{
Ying Taur
}

