% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{short_number}
\alias{short_number}
\title{Short number formatting}
\usage{
short_number(x, sig.digits = 2)
}
\arguments{
\item{x}{numeric vector to be formatted}

\item{sig.digits}{number of signficant digits to use, if the number is abbreviated}
}
\value{
character vector of formatted numbers
}
\description{
Use to abbreviate large numbers (e.g. 3450000 is '3.4M')
}
\examples{
short_number(3.4*10^(-1:10))
}
