% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{survival.frame.atrisk.table}
\alias{survival.frame.atrisk.table}
\title{At-Risk Table from Survival Data Frame}
\usage{
survival.frame.atrisk.table(
  t.breaks,
  sf,
  minus.epsilon.last.t = TRUE,
  melt = FALSE,
  row.name.xloc
)
}
\arguments{
\item{t.breaks}{vector of times to calculate at-risk}

\item{sf}{survival frame}

\item{minus.epsilon.last.t}{means we substract a small amt from last timepoint, because otherwise it's all NA.}
}
\description{
At-Risk Table from Survival Data Frame
}
\author{
Ying Taur
}
