% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yingtools2.R
\name{survival.frame.info}
\alias{survival.frame.info}
\title{Survival Frame Info}
\usage{
survival.frame.info(t, sf, infotype)
}
\arguments{
\item{t}{time}

\item{sf}{survival frame}

\item{infotype}{type of info needed. Needs to be a variable in survival frame.}
}
\value{
info needed from survival frame
}
\description{
Given survival.frame, and time, provide info in the survivalframe.
}
\author{
Ying Taur
}
