% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{tax.plot}
\alias{tax.plot}
\title{Plot tax}
\usage{
tax.plot(
  t,
  xvar = "sample",
  pctseqs = "pctseqs",
  unitvar = "Species",
  label = "Species",
  tax.levels = c("Superkingdom", "Phylum", "Class", "Order", "Family", "Genus",
    "Species"),
  data = FALSE,
  label.pct.cutoff = 0.3
)
}
\arguments{
\item{t}{data frame containing melted tax data. Needs to have vars sample, pctseqs, Kingdom, ... , Species}

\item{xvar}{xvar by which to plot data}

\item{pctseqs}{the relative abundance column name  (default \code{pctseqs})}

\item{unitvar}{unit var (default \code{Species})}

\item{label}{column name for labels var (default \code{Species})}

\item{tax.levels}{tax ranks.}

\item{data}{whether to return data frame}

\item{label.pct.cutoff}{cutoff by which to label abundances, stored in tax.label}
}
\value{
either ggplot2 object, or data frame.
}
\description{
Plot tax
}
\author{
Ying Taur
}
