% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{tax.plot}
\alias{tax.plot}
\title{Plot tax}
\usage{
tax.plot(t, xvar = "sample", data = FALSE, label.pct.cutoff = 0.3,
  use.cid.colors = TRUE)
}
\arguments{
\item{t}{data frame containing melted tax data. Needs to have vars sample, pctseqs, Kingdom, ... , Species}

\item{xvar}{xvar by which to plot data}

\item{data}{whether to return data frame}

\item{label.pct.cutoff}{cutoff by which to label abundances, stored in tax.label}

\item{cid.colors}{whether to use conventional cid colors.}
}
\value{
either ggplot2 object, or data frame.
}
\description{
Plot tax
}
\author{
Ying Taur
}
