% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{taxnonomy_scale}
\alias{taxnonomy_scale}
\title{Taxonomy scale constructor}
\usage{
taxnonomy_scale(
  aesthetic,
  values = NULL,
  breaks = waiver(),
  ...,
  limits = NULL
)
}
\description{
Similar to  \link[ggplot2:::manual_scale]{ggplot2::::manual_scale}
}
