% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiota2.R
\name{taxonomy_is_distinct}
\alias{taxonomy_is_distinct}
\title{Check if taxonomy levels are distinct.}
\usage{
taxonomy_is_distinct(
  data,
  taxranks = c("Superkingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")
)
}
\arguments{
\item{data}{tax data to be tested. Can be a \code{\link[phyloseq:phyloseq-class]{phyloseq}}, tax table (from \code{\link[=get.tax]{get.tax()}}), or otu-melt table (from \code{\link[=get.otu.melt]{get.otu.melt()}}).}

\item{taxranks}{character vector of tax levels to be tested.}
}
\value{
logical indicating whether or not taxonomy levels are distinct.
}
\description{
Check if taxonomy levels are distinct.
}
